### 2019-04-14 #1384 @banesullivan
Whoops, forgot to merge the paper into the `master` branch... on it.
