### 2019-04-14 674b12b
JOSS Paper (#40)
* Add draft of paper

* Finalize JOSS paper

* Edit paper
