+++ joss/paper.md
@@ -1,0 +1,154 @@
+---
+title: 'PVGeo: an open-source Python package for geoscientific visualization in VTK and ParaView'
+tags:
+  - Python
+  - visualization
+  - 3D
+  - geoscience
+authors:
+  - name: C. Bane Sullivan
+    orcid: 0000-0001-8628-4566
+    affiliation: 1
+  - name: Whitney J. Trainor-Guitton
+    orcid: 0000-0002-5726-3886
+    affiliation: 1
+affiliations:
+ - name: Department of Geophysics, Colorado School of Mines, Golden, CO, USA
+   index: 1
+
+date: 18 February 2019
+bibliography: paper.bib
+---
+
+# Summary
+
+PVGeo is an open-source Python package for geoscientific visualization and
+analysis, harnessing an already powerful software platform: the Visualization
+Toolkit (VTK) and its front-end application, ParaView.
+The VTK software platform is well-maintained, contains an expansive set of
+native functionality, and provides a robust foundation for scientific
+visualization, yet the development of tools compatible with geoscience data and
+models has been limited.
+As a software extension package to VTK and ParaView, PVGeo addresses the lack of
+geoscientific compatibility by creating a framework for geovisualization.
+PVGeo aims to make the process of importing geoscience data into VTK based
+software fluid and straightforward for users while providing a framework for new features
+avoiding the typical, ambitious programming endeavor of building VTK software
+plugins.
+We have developed this code library, PVGeo, to link geoscientific data and
+models with VTK-based 3D rendering environments like ParaView: an open-source
+platform built on top of VTK [@pvguide].
+Since VTK is an established and robust visualization platform, it provides
+a rich toolbox of features common for visualization and spatial analysis across
+disciplines [@vtkbook].
+Examples of standard features include volume rendering, glyphing, subsetting,
+K-Means clustering, volume interpolation, iso-contouring, and Virtual Reality
+[@pvguide], [@vtkbook].
+By linking geoscience to VTK and ParaView, geoscientists can harness all of the
+native tools within ParaView, and other VTK powered libraries like ParaViewWeb
+[@pvweb], VTK.js [@vtkjs], and `vtki` [@vtki] or extend that data into new
+domains like Virtual Reality, as outlined in Figure 1.
+PVGeo couples geoscientific information to software libraries at the forefront
+of scientific visualization, which enables scientists to cost-effectively and
+reproducibly communicate their findings.
+
+
+![Figure 1](images/expansion-diagram.png)
+**Figure 1**. PVGeo providing a link for geoscience to the VTK and ParaView
+realm of data visualization.
+
+
+
+## Background
+
+
+The results of geophysical imaging techniques often hold high significance to
+stakeholders yet the effective perception of those results remains a dynamic
+challenge.
+In the geosciences and especially the field of geophysics, researchers often
+need 3D and 4D (time-varying) visualizations to understand complex spatial and
+temporal relationships in data which are challenging to capture in 2D
+visualizations [@witter].
+Better perceptions or new understandings may arise from data when referenced in
+relation to intuitive features like topography, well locations, survey points,
+or other known information.
+Through these spatial relations, geoscientists and stakeholders can directly
+engage with their data to gain insight and begin to rapidly evaluate data and
+models either on various 2D planes simultaneously or in a complex 3D environment
+[@witter, @Carr1997].
+
+
+Geoscientists often use specific visualization software for different data
+processing routines which can lead to using several different visualization
+environments for a single project.
+Unfortunately, geoscientists are often left without a toolset for visual
+integration across those different data types, like pairing well locations,
+resource models, and geological in Figure 2, which can hinder their ability to
+interpret the spatial relationships of varying data types.
+
+
+![Figure 2](./images/forge-iso.png)
[…]
