### 2019-04-15 #1384 @banesullivan
Please note that there is a citation for `vtki` in this paper: #1382
