### 2019-04-19 62d03f3
Fix caption formatting for JOSS paper
The caption should be specified inside the `[]` when inserting an image.
