### 2019-04-19 e8582e7
Install NumPy before discretize on AppVeyor
