+++ appveyor.yml
@@ -32,6 +32,8 @@ install:
   - conda info -a
   # Upgrade pip
   - python -m pip install --upgrade pip
+  # Install NumPy ahead of time for discretize
+  - python -m pip install --upgrade numpy
   # Install extra dependencies
   - conda install -c conda-forge pyproj
   # Handle VTK if archetecture requires conda, else let setup.py install via pip
