### 2019-04-19 3839b62
JOSS: remove link inside figure caption
