+++ joss/paper.md
@@ -87,8 +87,7 @@ interpret the spatial relationships of varying data types.
 
 ![Demonstration of how PVGeo links various geoscientific
 data formats and filtering algorithms with VTK-based visualization software
-to create an integrated rendering environment
-([FORGE Geothermal Project](http://forge.pvgeo.org)).
+to create an integrated rendering environment (FORGE Geothermal Project).
 This rendering includes a digital land surface with overlain satellite imagery
 and geologic map, a subsurface temperature model, scattered points of the
 sampled temperature values, geophysical well logging data, GIS site boundary,
