### 2019-04-19 #1384 @banesullivan
How would I go about making sure Figure 2 isn't forced onto its own page? Ideally, I'd like it to be at the top of that page and then have the `Mentions` section start right under it.

I'm not familiar with the markdown to LaTeX converter used here and how I might implement a fix for this.
