### 2019-04-19 ef3738c
Add Travis test for Python 3.7
