### 2019-04-19 2eca407
Update Travis build to for Python 3.7
