+++ .travis.yml
@@ -1,11 +1,12 @@
-language: python
+sudo: required
+dist: xenial
 os:
-    - linux
+  - linux
+language: python
 python:
-    - 2.7 # build on Python 2 first
-    - 3.6
-    - 3.7
-sudo: false
+  - '3.5'
+  - '3.6'
+  - '3.7'
 install:
     - pip install -r requirements.txt
 script:
