### 2019-04-19 7cb5be2
Re-add Python 2.7 to Travis
