+++ .travis.yml
@@ -4,6 +4,7 @@ os:
   - linux
 language: python
 python:
+  - '2.7'
   - '3.5'
   - '3.6'
   - '3.7'
