### 2019-04-19 dcf207d
Fix python_requires in setup.py
