### 2019-04-19 23549a2
JOSS: Specify compatible Python versions (#43)
* Add python_requires to setup.py for #41

* Update Python version in README per #41

* Add AppVeyor test for Python 3.7

* Add Travis test for Python 3.7

* Update Travis build to for Python 3.7

* Re-add Python 2.7 to Travis

* Fix python_requires in setup.py
