+++ appveyor.yml
@@ -12,6 +12,9 @@ environment:
     - PYTHON_VERSION: 3.6
       CONDA: C:\Miniconda36
 
+    - PYTHON_VERSION: 3.7
+      CONDA: C:\Miniconda36
+
 image: Visual Studio 2015
 
 platform:
