### 2019-04-21 be5caaf
Use example gallery in autogenerated docs
