+++ docs/source/conf.py
@@ -16,6 +16,10 @@ import os
 import sys
 import shutil
 
+if sys.version_info >= (3, 0):
+    import faulthandler
+    faulthandler.enable()
+
 # Add PVGeo to the path
 path = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 print(path)
@@ -46,11 +50,34 @@ Generator().DocumentPackages([PVGeo, pvmacros],
             notify=False,
             )
 
+# HACK: Now append the examples toc to the end of the index
+example_toc = """
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Examples
+   :hidden:
+
+   examples/index
+
+"""
+with open(os.path.join(os.path.dirname(__file__), 'index.rst'), 'a') as f:
+    f.write(example_toc)
+
+import vtki
+import numpy as np
+# Manage errors
+vtki.set_error_output_file('errors.txt')
+# Ensure that offscreen rendering is used for docs generation
+vtki.OFF_SCREEN = True # Not necessary - simply an insurance policy
+# Preferred plotting style for documentation
+vtki.set_plot_theme('document')
+
 
 # -- Project information -----------------------------------------------------
 
 project = 'PVGeo'
-copyright = u'2018, Bane Sullivan, http:://banesullivan.com'
+copyright = u'2018-2019, Bane Sullivan, http:://banesullivan.com'
 author = 'Bane Sullivan'
 html_show_copyright = False
 html_show_sphinx = False
@@ -80,6 +107,8 @@ extensions = [
     'sphinx.ext.viewcode',
     'sphinx.ext.githubpages',
     'sphinxcontrib.napoleon',
+    'sphinx_copybutton',
+    'sphinx_gallery.gen_gallery'
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -213,3 +242,29 @@ intersphinx_mapping = {'https://docs.python.org/': None}
 
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = True
+
+
+
+# -- Sphinx Gallery Options
+from sphinx_gallery.sorting import FileNameSortKey
+
+sphinx_gallery_conf = {
+    # path to your examples scripts
+    "examples_dirs": [
+        "../../examples/",
+    ],
+    # path where to save gallery generated examples
+    "gallery_dirs": ["examples"],
+    # Patter to search for example files
+    "filename_pattern": r"\.py",
+    # Remove the "Download all examples" button from the top level gallery
+    "download_all_examples": False,
+    # Sort gallery example by file name instead of number of lines (default)
+    "within_subsection_order": FileNameSortKey,
+    # directory where function granular galleries are stored
+    "backreferences_dir": False,
+    # Modules for which function level galleries are created.  In
+    "doc_module": "PVGeo",
+    "image_scrapers": (vtki.Scraper(), 'matplotlib'),
+    "thumbnail_size": (350, 350),
+}
