+++ examples/README.txt
@@ -1,0 +1,6 @@
+.. _ref_examples:
+
+Quick Examples
+==============
+
+Here is a gallery of all the quick examples demonstrating what ``PVGeo`` can do!
