+++ docs/requirements.txt
@@ -18,3 +18,6 @@ pyproj>=1.9
 omf>=0.9.3
 omfvtk>=0.0.3
 espatools>=0.0.8
+doctr
+sphinx-copybutton
+sphinx-gallery
