### 2019-04-21 3202ae4
Fix travis docs build
