+++ docs/requirements.txt
@@ -6,18 +6,6 @@ sphinx-autobuild
 sphinx-rtd-theme
 sphinxcontrib-napoleon
 sphinxcontrib-websupport
-# From root level requirements:
-numpy>=1.13
-scipy>=1.1
-colour-runner
-codecov
-vtk>=8.1.0
-vtki>=0.15.0
-pandas>=0.23.4
-pyproj>=1.9
-omf>=0.9.3
-omfvtk>=0.0.3
-espatools>=0.0.8
 doctr
 sphinx-copybutton
 sphinx-gallery
