+++ docs/source/about-examples.rst
@@ -1,0 +1,194 @@
+About Examples
+==============
+
+`PVGeo` is deployed in various sub-packages called *suites*. These *suites*
+consist of a set of reader, filter, source, or writer algorithms (or any
+combination of those) for a general area of geoscientific processing and
+visualization.
+The following sections on this page demonstrate general procedures and syntax to
+use each type of algorithm within ParaView or directly in a Python environment.
+
+Take a look at the :ref:`gallery` for an outline of all the available
+examples at this time.
+
+If you think there may be a serious problem with an example, please open an
+issue on the `issues page`_ so that we can promptly fix it.
+
+.. _issues page: https://github.com/OpenGeoVis/PVGeo/issues
+
+
+Typical Usage
+-------------
+
+All algorithms deployed in *PVGeo* are useable in the following manners in a
+Python environment where the algorithm can be called and instantiated with
+keyword arguments for its parameters and then applied on some input data set.
+
+.. code-block:: python
+
+    import PVGeo
+    # PSEUDOCODE: Typical use of a PVGeo algorithm:
+    output = PVGeo.suite.Algorithm(**kwargs).apply(input)
+
+
+Or we can instantiate the algorithm for repetitive calls if, for example, we
+need to request varying time steps.
+
+.. code-block:: python
+
+    import PVGeo
+    # PSEUDOCODE: Typical use of a PVGeo algorithm:
+    alg = PVGeo.suite.Algorithm(**kwargs)
+    # Grab the output data object
+    output = alg.apply(input)
+    # Update the output to a desired time step
+    alg.UpdateTimeStep(6.0)
+
+
+Reader Algorithms
+-----------------
+
+A reader takes data from files and puts them into the proper VTK
+data structures so that we can visualize that data on the VTK or ParaView
+pipeline.
+ParaView and `vtki` come with a plethora of native data format readers but
+there are still many more formats in the geosciences that have not been
+implemented. By creating formats for common geoscientific formats, we hope to
+make the process of getting data into the ParaView pipeline or into `vtki`
+data structures as simple as possible.
+
+
+The file readers in *PVGeo* are available for use in the same manner as all
+algorithms in *PVGeo*. Readers are typically used in a manner that allows the
+reader algorithm to be repetitively called to request various time steps:
+
+.. code-block:: python
+
+    import PVGeo
+    # PSEUDOCODE: Typical use of a PVGeo reader:
+    reader = PVGeo.suite.Reader(**kwargs)
+    reader.AddFileName(['file%.2d' % i for i in range(20)])
+
+    # Grab the output data object
+    output = reader.apply() # NOTE: Readers have no input for the `apply()` call
+
+    # Update the output to a desired time step
+    reader.UpdateTimeStep(6.0)
+
+
+It is worth noting that if you have only one file (one time step) to read, then
+readers can be used to immediately produce a data object:
+
+.. code-block:: python
+
+    import PVGeo
+    # PSEUDOCODE: Typical use of a PVGeo reader:
+    output = PVGeo.suite.Reader(**kwargs).apply('fname.txt')
+
+
+
+
+Filter Algorithms
+-----------------
+
+A filter modifies, transforms, combines, analyses, processes, etc. data in VTK
+data structures on either a VTK or ParaView pipeline. Filters provide a means
+for changing how we visualize data or create a means of generating topology for
+an input data source to better represent that data in a 3D rendering environment.
+
+For example, we have developed a filter called Voxelize Points
+which takes a set of scattered points sampled on a rectilinear reference frame
+and generates voxels for every point such that the volume of data made by the
+points is filled with topologically connected cells.
[…]
