### 2019-04-21 c36e935
Fix typo in Surfer example
