### 2019-04-21 ff2ffb5
Comment out example code that doesn't run on Travis
