### 2019-04-21 2c2862c
A whole lot more pep8 refactoring
