+++ PVPlugins/PVGeo_UBC_Tools.py
@@ -40,22 +40,22 @@ class PVGeoTensorMeshReader(TensorMeshReader):
             panel_visibility="advanced"
             name="MeshFile"
             label="File Name Mesh"
-            command="SetMeshFileName"
+            command="set_mesh_filename"
             animateable="1"
             clean_command="ClearMesh"
             number_of_elements="1">
             <FileListDomain name="meshfile"/>
             <Documentation>This is the mesh file for a 2D or 3D UBC Mesh grid. This plugin only allows ONE mesh to be defined.</Documentation>
         </StringVectorProperty>''')
-    def SetMeshFileName(self, filename):
-        TensorMeshReader.SetMeshFileName(self, filename)
+    def set_mesh_filename(self, filename):
+        TensorMeshReader.set_mesh_filename(self, filename)
 
     @smproperty.xml('''
         <StringVectorProperty
           panel_visibility="default"
           name="ModelFiles"
           label="File Name(s) Model"
-          command="AddModelFileName"
+          command="add_model_file_name"
           animateable="1"
           repeat_command="1"
           clean_command="ClearModels"
@@ -63,9 +63,9 @@ class PVGeoTensorMeshReader(TensorMeshReader):
           <FileListDomain name="model_files"/>
           <Documentation>This is for a single sets of model files to append to the mesh as data time varying attributes. You can chose as many files as you would like for this for the given attribute.</Documentation>
         </StringVectorProperty>''')
-    def AddModelFileName(self, filename):
+    def add_model_file_name(self, filename):
         """Use to set the file names for the reader. Handles singlt string or list of strings."""
-        TensorMeshReader.AddModelFileName(self, filename)
+        TensorMeshReader.add_model_file_name(self, filename)
 
 
     @smproperty.doublevector(name="TimeDelta", default_values=1.0, panel_visibility="advanced")
@@ -77,14 +77,14 @@ class PVGeoTensorMeshReader(TensorMeshReader):
         """This is critical for registering the timesteps"""
         return TensorMeshReader.get_time_step_values(self)
 
-    @smproperty.xml(_helpers.get_property_xml(name='Use file as data name', command='SetUseFileName', default_values=True, help='A boolean to override the DataName and use model file name as data name.',
+    @smproperty.xml(_helpers.get_property_xml(name='Use file as data name', command='set_use_filename', default_values=True, help='A boolean to override the DataName and use model file name as data name.',
     panel_visibility="advanced"))
-    def SetUseFileName(self, flag):
-        TensorMeshReader.SetUseFileName(self, flag)
+    def set_use_filename(self, flag):
+        TensorMeshReader.set_use_filename(self, flag)
 
     @smproperty.stringvector(name='DataName', default_values='Data', panel_visibility="advanced")
-    def SetDataName(self, name):
-        TensorMeshReader.SetDataName(self, name)
+    def set_data_name(self, name):
+        TensorMeshReader.set_data_name(self, name)
 
 @smproxy.filter(name="PVGeoTensorMeshAppender",
        label=TensorMeshAppender.__displayname__)
@@ -102,7 +102,7 @@ class PVGeoTensorMeshAppender(TensorMeshAppender):
           panel_visibility="default"
           name="ModelFiles"
           label="File Name(s) Model"
-          command="AddModelFileName"
+          command="add_model_file_name"
           animateable="1"
           repeat_command="1"
           clean_command="ClearModels"
@@ -110,18 +110,18 @@ class PVGeoTensorMeshAppender(TensorMeshAppender):
           <FileListDomain name="model_files"/>
[…]
