### 2019-04-21 5a8e92c
Better plotting in SGeMS example
