+++ .travis.yml
@@ -31,10 +31,9 @@ script:
     - coverage run tests/__main__.py
     - codecov --token=8da9d332-1618-450d-8654-09a8ecf3c2dc
     - python tests/__main__.py close # exit with error code if problem occurs
-    # Make and deploy docs: doctr deploy --built-docs ./build/html/ .;
+    # Make and deploy docs TODO: set -e; <><><> doctr deploy --built-docs ./build/html/ .;
     - if [[($TRAVIS_PYTHON_VERSION == 3.7)]]; then
         echo "Building and uploading the docs";
-        set -e;
         cd ./docs;
         pip install -r ./requirements.txt;
         make html;
