### 2019-04-21 1494130
Make sure segfualt handler is enabled
