### 2019-04-21 5e2c01b
Make docs twice to avoid segfault issue
