### 2019-04-21 cb0851c
Fix requirements for docs
