+++ .travis.yml
@@ -26,7 +26,7 @@ before_install:
 install:
     - pip install -r requirements.txt
     - if [[($TRAVIS_PYTHON_VERSION == 3.7)]]; then
-        pip install -r ./requirements.txt;
+        pip install -r ./docs/requirements.txt;
       fi
     - python setup.py install
 
