+++ docs/source/dev-guide/resources.rst
@@ -1,0 +1,34 @@
+External References
+===================
+
+Here we have compiled a list of resources and reference material for learning
+how to use VTK in a Python environment and learning the ParaView interface.
+
+
+
+VTK & Python
+------------
+
+First, brush up on `vtki <http://docs.vtki.org`_. ``vtki`` provides a major
+foundation for PVGeo.
+
+
+- `Improved VTK – numpy integration | Kitware Blog <https://blog.kitware.com/improved-vtk-numpy-integration/>`_
+- `IPython Notebook & VTK | PyScience <https://pyscience.wordpress.com/2014/09/03/ipython-notebook-vtk/>`_
+- `A VTK pipeline primer (part 1) | Kitware Blog <https://blog.kitware.com/a-vtk-pipeline-primer-part-1/>`_
+- `vtkPythonAlgorithm is great | Kitware Blog <https://blog.kitware.com/vtkpythonalgorithm-is-great/>`_
+- `Developing HDF5 readers using vtkPythonAlgorithm | Kitware Blog <https://blog.kitware.com/developing-hdf5-readers-using-vtkpythonalgorithm/>`_
+- `VTK/Tutorials/New Pipeline - KitwarePublic <https://www.vtk.org/Wiki/VTK/Tutorials/New_Pipeline>`_
+- `VTK Python Examples <https://lorensen.github.io/VTKExamples/site/Python/>`_
+- `Create virtual environments for python with conda <https://uoa-eresearch.github.io/eresearch-cookbook/recipe/2014/11/20/conda/>`_
+- `ParaView/Simple ParaView 3 Python Filters - KitwarePublic <https://www.paraview.org/Wiki/ParaView/Simple_ParaView_3_Python_Filters>`_
+- `Streaming in VTK : Time | The Kitware Blog <https://blog.kitware.com/streaming-in-vtk-time/>`_
+
+
+ParaView
+--------
+
+- `The ParaView Guide <https://www.paraview.org/paraview-guide/>`_
+- `ParaViewWeb: Using ParaView’s Visualization and Data Analysis Capabilities within Web Applications | Kitware Blog <https://blog.kitware.com/paraviewweb-using-paraviews-visualization-and-data-analysis-capabilities-within-web-applications/>`_
+- `Proxy Hints And Annotations <https://www.paraview.org/ParaView/Doc/Nightly/www/cxx-doc/ProxyHints.html>`_
+- `ServerManager XML Hints <https://www.paraview.org/Wiki/ServerManager_XML_Hints>`_
