+++ docs/source/dev-guide/contributing.rst
@@ -1,0 +1,73 @@
+.. _Banes website: http://banesullivan.com
+.. _issues page: https://github.com/OpenGeoVis/PVGeo/issues
+.. _license: https://github.com/OpenGeoVis/PVGeo/blob/master/LICENSE
+
+Contributing
+============
+
+Thanks for taking the time to contribute!
+
+We absolutely welcome contributions and we hope that this guide will facilitate
+an understanding of the *PVGeo* code repository. It is important to note that
+the *PVGeo* software package is maintained on a volunteer basis and thus we need
+to foster a community that can support user questions and develop new features
+to make this software a useful tool for geoscientists.
+
+This page is dedicated to outline where you should start with your question,
+concern, feature request, or desire to contribute.
+
+Questions
+---------
+
+For general questions about the project, its applications, or about the
+non-technical, please do not create an issue but join us on
+[slack](http://slack.pvgeo.org) or send one of the developers an email.
+The current project maintainer is Bane Sullivan, feel
+free to inquire with Bane through `Banes website`_.
+
+For more technical questions, you are welcome to create an issue on the
+`issues page`_ with a *question* label which we will try to address.
+Through posting on the issues page, your question can be addressed by community
+members with the needed expertise and the information gained will remain
+available on the issues page for other users.
+
+
+Reporting Bugs
+--------------
+
+If you stumble across any bugs, crashes, or concerning quirks while using code
+distributed here, please report it on the `issues page`_ with an
+appropriate label so we can promptly address it.
+When reporting an issue, please be overly descriptive so that we may reproduce it.
+Whenever possible, please provide tracebacks, screenshots, and sample files
+to help us address the issue.
+
+Feature Requests
+----------------
+
+We encourage users to submit ideas for improvements to *PVGeo* code base!
+Please create an issue on the `issues page`_ with a *Feature Request*
+label to suggest an improvement.
+Please use a descriptive title and provide ample background information to help
+the community implement that functionality. For example, if you would like a
+reader for a specific file format, please provide a link to documentation of
+that file format and possibly provide some sample files with screenshots to work
+with. We will use the issue thread as a place to discuss and provide feedback.
+
+Contributing New Code
+---------------------
+
+If you have an idea for how to improve *PVGeo*, please first create an issue as
+a feature request which we can use as a discussion thread to work through how to
+implement the contribution.
+
+Once you are ready to start coding and develop for *PVGeo*, please take a look
+at the remainder of the pages in this Development Guide.
+
+Licensing
+---------
+
+All contributed code will be licensed under *PVGeos'* `license`_.
+If you did not write the code yourself, it is your responsibility to ensure that
+the existing license is compatible and included in the contributed files or you
+can obtain permission from the original author to relicense the code.
