+++ pvmacros/export/__init__.py
@@ -1,3 +1,160 @@
+"""
+Export
+======
+
+On this page, we demonstrate how to quickly share a 3D rendering of your
+ParaView visualizations with anyone who has access to the internet so that that
+can explore the whole scene in a dynamic manner.
+
+Motivation
+----------
+In order to effectively communicate our geoscientific findings, we often need to
+share our 3D visualizations with interested stakeholders. These interested
+parties are likely not going to have ParaView or other visualization software
+at hand. Thus we desire to have a means to export our complex visualizations in
+ParaView to a simple, shareable format that anyone can view. To accomplish this,
+we will take advantage of vtk.js and its standalone web viewer for vtk.js
+formats.
+
+Would not it be great to send your client or interested parties an interactive
+3D scene of your Geophysical findings like the example below?
+
+.. raw:: html
+
+    <div style="position: relative; padding-bottom: 56.25%; height: 0; overflow: hidden; max-width: 100%; height: auto;">
+            <iframe src="http://tunnels.pvgeo.org/" frameborder="0" allowfullscreen style="position: absolute; top: 0; left: 0; width: 100%; height: 100%;"></iframe>
+    </div>
+
+
+
+VTK.js
+------
+
+`vtk.js <https://kitware.github.io/vtk-js/>`_ is a rendering library made for
+scientific visualization on the web. This code base brings high performance
+rendering into anyone's web browser. This library allows us to export complex
+scenes from ParaView and share them with anyone that has a web browser like
+Safari or Google Chrome.
+
+The vtk.js library has an open-source `standalone scene viewer <https://kitware.github.io/vtk-js/examples/StandaloneSceneLoader/StandaloneSceneLoader.html>`_
+which they have a `nice demo <https://kitware.github.io/vtk-js/examples/StandaloneSceneLoader.html>`_.
+The first link can either be downloaded as an HTML file to be ran locally, or
+you can go to that link and run from the vtk.js server. vtk.js also published a
+scene export macro for ParaView that compresses a data scene in ParaView to a
+shareable format for viewing on the web. The `macro from the vtk.js library <https://raw.githubusercontent.com/Kitware/vtk-js/master/Utilities/ParaView/export-scene-macro.py>`_
+can be used but we also deploy an updated (we think more robust) version of
+this export macro in the sub-module `export` of our Python module `pvmacros`.
+
+Demo Shareable Format
++++++++++++++++++++++
+
+Here are some samples to demonstrate the web viewer which we host on
+`viewer.vtki.org <http://viewer.vtki.org>`_. We have included a few of our
+scenes and one of the vtk.js sample scenes for you to demo:
+
+- `Fluvial Channels <http://viewer.vtki.org/?fileURL=https://dl.dropbox.com/s/qnahdwedjwndo7t/fluvsim_channels.vtkjs?dl=0>`_
+- `Volcano <http://volcano.pvgeo.org>`_
+- `Ripple <http://ripple.pvgeo.org>`_
+- `Tunnels <http://tunnels.pvgeo.org>`_
+- `vtk.js Sample Scene <http://viewer.vtki.org/?fileURL=https://data.kitware.com/api/v1/file/587003c38d777f05f44a5c93/download>`_
+
+
+
+Example Use
+-----------
+
+First, make a complex scene in ParaView that you might like to share with someone.
+Now that you have your scene loaded, open the python shell from'View->Python Shell'
+(or 'Tools->Python Shell' depending on your ParaView version) within ParaView.
+From here, import our Python module delivered in the repository called
+``#!py pvmacros``. From the ``#!py export`` sub-module, there is a function called
+``#!py def exportVTKjs()`` which takes two optional arguments (`FileName` string
[…]
