+++ docs/source/dev-guide/resources.rst
@@ -9,7 +9,7 @@ how to use VTK in a Python environment and learning the ParaView interface.
 VTK & Python
 ------------
 
-First, brush up on `vtki <http://docs.vtki.org`_. ``vtki`` provides a major
+First, brush up on `vtki <http://docs.vtki.org>`_. ``vtki`` provides a major
 foundation for PVGeo.
 
 
