+++ docs/source/overview/getting-started.rst
@@ -1,0 +1,261 @@
+Getting Started
+===============
+
+Using PVGeo in a Python Environment
+-----------------------------------
+
+If you'd like to use PVGeo in Python (>=3.6) without ParaView then simply
+install PVGeo to your active Python environment (please note that we do not
+support the use of PVGeo in a Python 2 environment unless it is directly in
+ParaView's build).
+
+Install PVGeo via `pip <https://pypi.org/project/PVGeo/>`_::
+
+    pip install PVGeo
+
+
+Optional dependencies
++++++++++++++++++++++
+
+PVGeo has a few non-required dependencies that enable more algorithms and
+features when available. All requirements can be found in the
+`requirements.txt <https://github.com/OpenGeoVis/PVGeo/blob/master/requirements.txt>`_
+file in the repo but the needed requirements for PVGeo to work will be installed
+with PVGeo. Some useful dependencies:
+
+- `discretize <https://pypi.org/project/discretize/>`_: Adds algorithms that harnesses ``discretize``'s finite volume code and file IO methods.
+- `pyproj <https://pypi.org/project/pyproj/>`_: Adds algorithms that can perform coordinate transformations
+
+
+
+Using PVGeo in ParaView
+-----------------------
+
+If you'd like to use PVGeo directly in ParaView's graphical user interface,
+you must follow the remaining steps in this section very carefully which set up
+an isolated **Python 2.7 environment** that will be shared with your installation of
+ParaView. Note that after you create this environment, you should leave it alone
+and install PVGeo using the steps in the section above for the Python
+environments you use everyday.
+
+
+A Brief Introduction to ParaView
+++++++++++++++++++++++++++++++++
+
+ParaView is an open-source platform that can visualize 2D, 3D, and 4D
+(time-varying) datasets. ParaView can process multiple very large data sets in
+parallel then later collect the results to yield a responsive graphics
+environment with which a user can interact. The better the processor and
+graphics hardware the machine or machines hosting the software, the faster
+and better ParaView will run. However, it can run quite well on a laptop
+with a standard graphics card such as a MacBook Pro.
+
+Since ParaView is an open source application, anyone can download the program
+and its source code for modifications. The easiest way to get started with
+ParaView is to `download the compiled binary installers for your operating system <https://www.paraview.org/download/>`_.
+
+For further help, check out the `documentation <https://www.paraview.org/documentation/>`_
+provided by Kitware. In particular, the two worth looking through for a quick
+tour of ParaView are the **The ParaView Guide** and **The ParaView Tutorial.**
+One is a tutorial of the ParaView software and shows the user how to create
+sources, apply filters, and more. The other is a guide on how to do scripting,
+macros, and more intense use of the application.
+
+Install ParaView
+++++++++++++++++
+
+Open the downloaded installer from
+`ParaView's website <https://www.paraview.org/download/>`_ for ParaView 5.6.x
+(or greater) and follow the prompts with the installer.
+
+Tour around software:
+Take a look at Section 2.1 of **The ParaView Tutorial** for details of the
+application’s GUI environment. Chapter 2 of the tutorial as a whole does an
+excellent job touring the software and its workflow for those unfamiliar with
+the software and its general capabilities.
+
+One convenient feature is to save the state of the ParaView environment.
+This saves all the options you selected for all the filters you applied to
+visualize some data. Select *File->Save State…* (*Note:* this saves the
[…]
