### 2019-04-21 f05cebb
pep8 refactoring (#38)
pep8 refactoring
