+++ PVGeo/filters/math.py
@@ -24,6 +24,15 @@ class ArrayMath(FilterPreserveTypeBase):
     perfrom math operations. The input arrays are used in their order of
     selection for the operations.
 
+    Args:
+        multiplier (float) : a static shifter/scale factor across the array
+            after normalization.
+
+        new_name (str): The new array's string name
+
+        operation (str, int, or callable): The operation as a string key, int
+            index, or callable method
+
     **Available Math Operations:**
 
     - `add`: This adds the two data arrays together
@@ -38,13 +47,13 @@ class ArrayMath(FilterPreserveTypeBase):
         FilterPreserveTypeBase.__init__(self)
         # Parameters:
         self.__multiplier = kwargs.get('multiplier', 1.0)
-        self.__newName = kwargs.get('newName', 'Mathed Up')
-        self.__inputArray1 = [None, None]
-        self.__inputArray2 = [None, None]
+        self.__new_name = kwargs.get('new_name', 'Mathed Up')
+        self.__input_array_1 = [None, None]
+        self.__input_array_2 = [None, None]
         # Convert operation to callable method
         op = kwargs.get('operation', 'add')
         if isinstance(op, (str, int)):
-            op = self.GetOperation(op)
+            op = self.get_operation(op)
         self.__operation = op
 
 
@@ -57,22 +66,26 @@ class ArrayMath(FilterPreserveTypeBase):
 
     @staticmethod
     def _multiply(arr1, arr2):
+        """Mutlipies two input NumPy arrays"""
         return arr1*arr2
 
     @staticmethod
     def _divide(arr1, arr2):
+        """Divides two input NumPy arrays"""
         return arr1/arr2
 
     @staticmethod
     def _add(arr1, arr2):
+        """Adds two input NumPy arrays"""
         return arr1+arr2
 
     @staticmethod
     def _subtract(arr1, arr2):
+        """Subtracts two input NumPy arrays"""
         return arr1-arr2
 
     @staticmethod
-    def GetOperations():
+    def get_operations():
         """Returns the math operation methods as callable objects in a
         dictionary
         """
@@ -86,29 +99,29 @@ class ArrayMath(FilterPreserveTypeBase):
         return ops
 
     @staticmethod
-    def GetOperationNames():
+    def get_operation_names():
         """Gets a list of the math operation keys
 
         Return:
             list(str): the keys for getting the math operations
         """
-        ops = ArrayMath.GetOperations()
+        ops = ArrayMath.get_operations()
         return list(ops.keys())
 
     @staticmethod
-    def GetOperation(idx):
+    def get_operation(idx):
         """Gets a math operation based on an index in the keys
 
         Return:
             callable: the math operation method
         """
         if isinstance(idx, str):
-            return ArrayMath.GetOperations()[idx]
-        n = ArrayMath.GetOperationNames()[idx]
-        return ArrayMath.GetOperations()[n]
+            return ArrayMath.get_operations()[idx]
+        n = ArrayMath.get_operation_names()[idx]
+        return ArrayMath.get_operations()[n]
 
 
-    def _MathUp(self, pdi, pdo):
+    def _math_up(self, pdi, pdo):
         """Make sure to pass array names and integer associated fields.
         Use helpers to get these properties.
         """
@@ -116,24 +129,24 @@ class ArrayMath(FilterPreserveTypeBase):
             # TODO: test this
             pdo = pdi.DeepCopy()
         # Get the input arrays
-        field1, name1 = self.__inputArray1[0], self.__inputArray1[1]
-        field2, name2 = self.__inputArray2[0], self.__inputArray2[1]
+        field1, name1 = self.__input_array_1[0], self.__input_array_1[1]
[…]
