+++ tests/gmggroup_test.py
@@ -1,10 +1,5 @@
 import unittest
 import os
-import numpy as np
-
-
-# VTK imports:
-from vtk.util import numpy_support as nps
 
 from PVGeo import _helpers
 
@@ -29,11 +24,11 @@ if omf_avail:
         """
 
         def setUp(self):
-            self.fname = os.path.join(os.path.dirname(__file__), 'data/test_file.omf')
+            self.filename = os.path.join(os.path.dirname(__file__), 'data/test_file.omf')
 
 
         def test_read_project(self):
             """`OMFReader`: read whole project file"""
-            proj = OMFReader().Apply(self.fname)
+            proj = OMFReader().apply(self.filename)
             self.assertIsNotNone(proj)
             self.assertEqual(proj.GetNumberOfBlocks(), 9)
