### 2019-04-21 40d4c6f
[docs]: use example gallery in autogenerated documentation (#44)
* Many pep8 changes

* Fix for * in range(len()) issues

* Fix type checking issues

* Fix trailing whitespace

* Fix except all issues

* Fix error prone issues

* Remove unused imports

* Fix redefinition issue

* Fix mutable arguments issue

* Fix redefinition of builtins

* Fix unused class methods

* Fix use of assert

* Major internal refactor of keyword arguments

* Ensure all of PVGeo has docstrings

* _helpers module refactored to PEP8

* filters module is mostly refactored to PEP8

* Use example gallery in autogenerated
[…]
