+++ _sources/examples/gslib/sg_execution_times.rst.txt
@@ -1,0 +1,10 @@
+
+:orphan:
+
+.. _sphx_glr_examples_gslib_sg_execution_times:
+
+Computation times
+=================
+**00:02.316** total execution time for **examples_gslib** files:
+
+- **00:02.316**: :ref:`sphx_glr_examples_gslib_read-gslib.py` (``read-gslib.py``)
