+++ _sources/content/pvmacros/vis/axes.rst.txt
@@ -1,0 +1,36 @@
+
+
+axes
+====
+
+.. automodule:: pvmacros.vis.axes
+    :private-members:
+
+
+
+Custom Axis Ticks
+-----------------
+
+.. autofunction:: pvmacros.vis.axes.customAxisTicks
+
+
+
+Reset Axis Ticks
+----------------
+
+.. autofunction:: pvmacros.vis.axes.resetAxisTicks
+
+
+
+Scale Axis
+----------
+
+.. autofunction:: pvmacros.vis.axes.scaleAxis
+
+
+
+Set Axis Labels from Bounds
+---------------------------
+
+.. autofunction:: pvmacros.vis.axes.setAxisLabelsFromBounds
+
