+++ _sources/examples/filters-general/add-cell-connectivity-to-points.rst.txt
@@ -122,7 +122,7 @@ Now it looks good (see image below)!
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.584 seconds)
+   **Total running time of the script:** ( 0 minutes  0.609 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_add-cell-connectivity-to-points.py:
