### 2019-04-21 b288006
More pep8 refactoring
