+++ PVGeo/ubc/tensor.py
@@ -100,7 +100,7 @@ class TensorMeshReader(ubcMeshReaderBase):
             model = model.flatten()
 
         # Convert data to VTK data structure and append to output
-        c = interface.convertArray(model, name=data_name, deep=True)
+        c = interface.convert_array(model, name=data_name, deep=True)
         # THIS IS CELL DATA! Add the model data to CELL data:
         mesh.GetCellData().AddArray(c)
         return mesh
@@ -149,11 +149,11 @@ class TensorMeshReader(ubcMeshReaderBase):
             c = np.array(c,dtype=float)
             if z:
                 c = -c[::-1]
-            return interface.convertArray(c,deep=True)
+            return interface.convert_array(c,deep=True)
 
         xcoords = _genCoords(xpts, xdisc)
         zcoords = _genCoords(zpts, zdisc, z=True)
-        ycoords = interface.convertArray(np.zeros(1),deep=True)
+        ycoords = interface.convert_array(np.zeros(1),deep=True)
 
         output.SetDimensions(nx,2,nz) # note this subtracts 1
         output.SetXCoordinates(xcoords)
@@ -202,16 +202,16 @@ class TensorMeshReader(ubcMeshReaderBase):
         """Helper method to read a 2D mesh
         """
         # Construct/read the mesh
-        if self.NeedToReadMesh():
+        if self.need_to_readMesh():
             TensorMeshReader.ubcMesh2D(filename_mesh, self.__mesh)
-            self.NeedToReadMesh(flag=False)
+            self.need_to_readMesh(flag=False)
         output.DeepCopy(self.__mesh)
-        if self.NeedToReadModels() and self.ThisHasModels():
+        if self.need_to_readModels() and self.this_has_models():
             self.__models = []
             for f in filename_models:
                 # Read the model data
                 self.__models.append(TensorMeshReader.ubcModel2D(f))
-            self.NeedToReadModels(flag=False)
+            self.need_to_readModels(flag=False)
         return output
 
 
@@ -286,9 +286,9 @@ class TensorMeshReader(ubcMeshReaderBase):
         # Set the dims and coordinates for the output
         output.SetDimensions(dim[0],dim[1],dim[2])
         # Convert to VTK array for setting coordinates
-        output.SetXCoordinates(interface.convertArray(cox,deep=True))
-        output.SetYCoordinates(interface.convertArray(coy,deep=True))
-        output.SetZCoordinates(interface.convertArray(coz,deep=True))
+        output.SetXCoordinates(interface.convert_array(cox,deep=True))
+        output.SetYCoordinates(interface.convert_array(coy,deep=True))
+        output.SetZCoordinates(interface.convert_array(coz,deep=True))
 
         return output
 
@@ -296,16 +296,16 @@ class TensorMeshReader(ubcMeshReaderBase):
     def __ubcMeshData3D(self, filename_mesh, filename_models, output):
         """Helper method to read a 3D mesh"""
         # Construct/read the mesh
-        if self.NeedToReadMesh():
+        if self.need_to_readMesh():
             TensorMeshReader.ubcMesh3D(filename_mesh, self.__mesh)
-            self.NeedToReadMesh(flag=False)
+            self.need_to_readMesh(flag=False)
         output.DeepCopy(self.__mesh)
-        if self.NeedToReadModels() and self.ThisHasModels():
+        if self.need_to_readModels() and self.this_has_models():
             self.__models = []
             for f in filename_models:
                 # Read the model data
                 self.__models.append(TensorMeshReader.ubcModel3D(f))
-            self.NeedToReadModels(flag=False)
+            self.need_to_readModels(flag=False)
         return output
 
[…]
