### 2019-04-21 740477b
Add UBC examples
