### 2019-04-21 5c34095
Update GSLib examples
