+++ _sources/examples/filters-general/array-math.rst.txt
@@ -72,7 +72,7 @@ Use the filter:
 
  .. code-block:: none
 
-    UniformGrid (0x7f7026508648)
+    UniformGrid (0x7f4ec109ba08)
       N Cells:      243
       N Points:     400
       X Bounds:     0.000e+00, 9.000e+00
@@ -120,7 +120,7 @@ Use a custom math operation:
 
  .. code-block:: none
 
-    UniformGrid (0x7f70265088e8)
+    UniformGrid (0x7f4ec109bca8)
       N Cells:      243
       N Points:     400
       X Bounds:     0.000e+00, 9.000e+00
