### 2019-04-21 d0270b3
Remove disclaimer from docs homepage
