+++ docs/index_base.rst
@@ -51,12 +51,7 @@ These tools are tailored to data visualization in the geosciences with a heavy
 focus on structured data sets like 2D or 3D time-varying grids.
 
 This website hosts the code documentation for the ``PVGeo`` python package found
-on `GitHub`_ and `PyPI`_. This website strictly documents the code so that users
-have a convenient and familiar means of searching through the library to
-understand how to use PVGeo in a Python environment outside of ParaView.
-If you are searching for examples and demonstrations on how to use the ``PVGeo``
-library, then head over to the `full website`_ where you can find tutorials with
-sample data sets and links to many other helpful resources.
+on `GitHub`_ and `PyPI`_.
 
 For a quick overview of how ``PVGeo`` can be used in a Python environment or
 directly within ParaView, please checkout the code snippets and videos on the
