+++ _sources/examples/filters-general/array-math.rst.txt
@@ -72,7 +72,7 @@ Use the filter:
 
  .. code-block:: none
 
-    UniformGrid (0x7f4ec109ba08)
+    UniformGrid (0x7fe6d315a8e8)
       N Cells:      243
       N Points:     400
       X Bounds:     0.000e+00, 9.000e+00
@@ -120,7 +120,7 @@ Use a custom math operation:
 
  .. code-block:: none
 
-    UniformGrid (0x7f4ec109bca8)
+    UniformGrid (0x7fe6d315ab88)
       N Cells:      243
       N Points:     400
       X Bounds:     0.000e+00, 9.000e+00
@@ -143,7 +143,7 @@ Use a custom math operation:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.066 seconds)
+   **Total running time of the script:** ( 0 minutes  0.070 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_array-math.py:
