+++ _sources/examples/filters-general/many-slices-along-points.rst.txt
@@ -112,7 +112,7 @@ slices to what you prefer!
 
  .. code-block:: none
 
-    MultiBlock (0x7fe6d31056a8)
+    MultiBlock (0x7f39e209c6a8)
       N Blocks:     6
       X Bounds:     0.000, 250.000
       Y Bounds:     0.000, 250.000
@@ -140,7 +140,7 @@ slices to what you prefer!
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  3.890 seconds)
+   **Total running time of the script:** ( 0 minutes  3.760 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_many-slices-along-points.py:
