### 2019-04-21 1dd3a3e
Remove windows users alert from docs homepage
