+++ docs/index_base.rst
@@ -124,14 +124,6 @@ pip:
     pip install PVGeo
 
 
-
-.. warning::
-
-    Windows users: Please see installation instructions on `GitHub README`_.
-
-.. _GitHub README: https://github.com/OpenGeoVis/PVGeo/#getting-started
-
-
 Now ``PVGeo`` is ready for use in your standard python environment. To use the
 *PVGeo* library as plugins in `ParaView`_, please see the
 `detailed explanation here`_.
