+++ docs/source/dev-guide/snippets/index.rst
@@ -3,6 +3,5 @@ Development Example Snippets
 
 .. toctree::
    :maxdepth: 5
-   :caption: PVGeo:
 
    composite-data-writers
