+++ _sources/examples/filters-general/extract-array.rst.txt
@@ -57,12 +57,12 @@ Construct the filter
 
  .. code-block:: none
 
-    vtkTable (0x51c6330)
+    vtkTable (0x55748a0)
       Debug: Off
       Modified Time: 21889
       Reference Count: 2
       Registered Events: (none)
-      Information: 0x51750d0
+      Information: 0x4a72c30
       Data Released: False
       Global Release Data: Off
       UpdateTime: 21901
@@ -99,7 +99,7 @@ Construct the filter
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.072 seconds)
+   **Total running time of the script:** ( 0 minutes  0.073 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_extract-array.py:
