+++ _sources/examples/filters-general/points-to-tube.rst.txt
@@ -73,7 +73,7 @@ then builds a polygonal tube around that line with some specified radius and num
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.484 seconds)
+   **Total running time of the script:** ( 0 minutes  0.472 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_points-to-tube.py:
