+++ _sources/examples/filters-general/append-cell-centers.rst.txt
@@ -44,7 +44,7 @@ Use an example mesh from vtki
 
  .. code-block:: none
 
-    RectilinearGrid (0x7f3200f6a528)
+    RectilinearGrid (0x7fce83776e88)
       N Cells:      16146
       N Points:     18144
       X Bounds:     -3.500e+02, 1.350e+03
@@ -71,7 +71,7 @@ Run the PVGeo algorithm
 
  .. code-block:: none
 
-    RectilinearGrid (0x7f3200f6a5e8)
+    RectilinearGrid (0x7fce8378d2e8)
       N Cells:      16146
       N Points:     18144
       X Bounds:     -3.500e+02, 1.350e+03
@@ -96,7 +96,7 @@ Run the PVGeo algorithm
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.841 seconds)
+   **Total running time of the script:** ( 0 minutes  0.799 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_append-cell-centers.py:
