+++ _sources/examples/filters-general/voxelize-points.rst.txt
@@ -94,7 +94,7 @@ object (essentially a point cloud).
 
  .. code-block:: none
 
-    PolyData (0x7f3200ee18e8)
+    PolyData (0x7fce837095e8)
       N Cells:      499200
       N Points:     499200
       X Bounds:     3.268e+05, 3.302e+05
@@ -170,7 +170,7 @@ The full pipeline method
     Voxelizing... done.
     Recovered Angle (deg.): 53.550
     Recovered Cell Sizes: (25.00, 25.00, 25.00)
-    UnstructuredGrid (0x7f3200ee1ac8)
+    UnstructuredGrid (0x7fce83709828)
       N Cells:      499200
       N Points:     524064
       X Bounds:     3.268e+05, 3.302e+05
@@ -219,7 +219,7 @@ create 3 orthogonal slices through a data volume.
 
  .. code-block:: none
 
-    MultiBlock (0x7f3200ee17c8)
+    MultiBlock (0x7fce83709528)
       N Blocks:     3
       X Bounds:     326804.336, 330185.260
       Y Bounds:     4406253.954, 4409862.606
@@ -262,7 +262,7 @@ Now display the slices and clipped volume
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  29.726 seconds)
+   **Total running time of the script:** ( 0 minutes  27.479 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_voxelize-points.py:
