+++ _sources/examples/ubc/sg_execution_times.rst.txt
@@ -5,8 +5,8 @@
 
 Computation times
 =================
-**00:08.580** total execution time for **examples_ubc** files:
+**00:09.180** total execution time for **examples_ubc** files:
 
-- **00:04.017**: :ref:`sphx_glr_examples_ubc_read-topo.py` (``read-topo.py``)
-- **00:03.887**: :ref:`sphx_glr_examples_ubc_read-tensor-mesh.py` (``read-tensor-mesh.py``)
-- **00:00.675**: :ref:`sphx_glr_examples_ubc_read-grv-obs.py` (``read-grv-obs.py``)
+- **00:04.205**: :ref:`sphx_glr_examples_ubc_read-tensor-mesh.py` (``read-tensor-mesh.py``)
+- **00:04.164**: :ref:`sphx_glr_examples_ubc_read-topo.py` (``read-topo.py``)
+- **00:00.811**: :ref:`sphx_glr_examples_ubc_read-grv-obs.py` (``read-grv-obs.py``)
