+++ _sources/examples/gslib/sg_execution_times.rst.txt
@@ -5,7 +5,7 @@
 
 Computation times
 =================
-**00:48.297** total execution time for **examples_gslib** files:
+**00:48.058** total execution time for **examples_gslib** files:
 
-- **00:47.833**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
-- **00:00.463**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
+- **00:47.561**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
+- **00:00.497**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
