+++ _sources/content/pvmacros/vis/index.rst.txt
@@ -4,7 +4,7 @@ Visualization
 
 .. toctree::
    :maxdepth: 5
-   :hidden:
+
     
    
    axes.rst
