+++ _sources/content/PVGeo/_helpers/index.rst.txt
@@ -4,7 +4,7 @@ Internal Helpers
 
 .. toctree::
    :maxdepth: 5
-   :hidden:
+
     
    
    arrays.rst
