### 2019-04-21 66ac23d
[BUGFIX] AddCellConnToPoints filter
