+++ docs/index_base.rst
@@ -1,10 +1,6 @@
-.. PVGeo documentation master file, created by
-   sphinx-quickstart on Tue Jul 10 19:56:04 2018.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
+#############################
 Welcome to PVGeo's code docs!
-=============================
+#############################
 
 
 .. image:: https://img.shields.io/badge/demos-grey.svg?logo=vimeo
