+++ docs/requirements.txt
@@ -1,5 +1,5 @@
 # For building the documentation:
-gendocs>=0.3.5
+gendocs>=0.4.0
 mock
 Sphinx
 sphinx-autobuild
