### 2019-04-21 cad77c3
Revert default cell type for AddCellConnToPoints
