+++ _sources/examples/grids/extract-topography.rst.txt
@@ -201,7 +201,7 @@ Note that there are other extraction operations like an ``'intersection'``:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  9.465 seconds)
+   **Total running time of the script:** ( 0 minutes  10.089 seconds)
 
 
 .. _sphx_glr_download_examples_grids_extract-topography.py:
