+++ _sources/examples/filters-general/extract-array.rst.txt
@@ -57,18 +57,18 @@ Construct the filter
 
  .. code-block:: none
 
-    vtkTable (0x4a55fa0)
+    vtkTable (0x5418e80)
       Debug: Off
-      Modified Time: 21118
+      Modified Time: 21270
       Reference Count: 2
       Registered Events: (none)
-      Information: 0x4a51c90
+      Information: 0x58bd700
       Data Released: False
       Global Release Data: Off
-      UpdateTime: 21130
+      UpdateTime: 21282
       Field Data:
         Debug: Off
-        Modified Time: 21112
+        Modified Time: 21264
         Reference Count: 1
         Registered Events: (none)
         Number Of Arrays: 0
@@ -76,7 +76,7 @@ Construct the filter
         Number Of Tuples: 0
       RowData: 
         Debug: Off
-        Modified Time: 21126
+        Modified Time: 21278
         Reference Count: 1
         Registered Events: (none)
         Number Of Arrays: 1
@@ -99,7 +99,7 @@ Construct the filter
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.053 seconds)
+   **Total running time of the script:** ( 0 minutes  0.062 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_extract-array.py:
