### 2019-04-23 91e2423
[BUGFIX] AddCellConnToPoints sorting issue
