### 2019-04-23 c29aba4
Add ability to close loops for AddCellConnToPoints filter
