### 2019-04-23 3b36594
Default AddCellConnToPoints filter to not preserve vertice cells
