+++ tests/filters_test.py
@@ -605,8 +605,7 @@ class TestAddCellConnToPoints(TestBase):
         f.set_cell_type(vtk.VTK_POLY_LINE)
         f.Update()
         output = f.GetOutput()
-        # NOTE: the algorithm adds vertice and line cells
-        self.assertEqual(len(self.pts) + 1, output.GetNumberOfCells())
+        self.assertEqual(1, output.GetNumberOfCells())
         # Now test nearest neighbor functionality
         self.makeComplicatedInput()
         f = AddCellConnToPoints()
@@ -615,8 +614,7 @@ class TestAddCellConnToPoints(TestBase):
         f.set_use_nearest_nbr(True)
         f.Update()
         output = f.GetOutput()
-        # NOTE: the algorithm adds vertice and line cells
-        self.assertEqual(len(self.pts) + 1, output.GetNumberOfCells())
+        self.assertEqual(1, output.GetNumberOfCells())
         # Its fairly difficult to test the nearest neighbor approximations...
         # This was done visually
         # The above test is just there to make sure no errors are thrown
@@ -631,8 +629,7 @@ class TestAddCellConnToPoints(TestBase):
         f.set_cell_type(vtk.VTK_LINE)
         f.Update()
         output = f.GetOutput()
-        # NOTE: the algorithm adds vertice and line cells
-        self.assertEqual(len(self.pts) + len(self.pts)-1, output.GetNumberOfCells())
+        self.assertEqual(len(self.pts)-1, output.GetNumberOfCells())
         # Now test nearest neighbor functionality
         self.makeComplicatedInput()
         f = AddCellConnToPoints()
@@ -641,8 +638,7 @@ class TestAddCellConnToPoints(TestBase):
         f.set_use_nearest_nbr(True)
         f.Update()
         output = f.GetOutput()
-        # NOTE: the algorithm adds vertice and line cells
-        self.assertEqual(len(self.pts) + len(self.pts)-1, output.GetNumberOfCells())
+        self.assertEqual(len(self.pts)-1, output.GetNumberOfCells())
         # Its fairly difficult to test the nearest neighbor approximations...
         # This was done visually in ParaView.
         # The above test is just there to make sure no errors are thrown
@@ -657,8 +653,7 @@ class TestAddCellConnToPoints(TestBase):
         f.set_cell_type(vtk.VTK_LINE)
         f.Update()
         output = f.GetOutput()
-        # NOTE: the algorithm adds vertice and line cells
-        self.assertEqual(len(self.pts) + len(self.pts), output.GetNumberOfCells())
+        self.assertEqual(len(self.pts), output.GetNumberOfCells())
         # Now test nearest neighbor functionality
         self.makeComplicatedInput()
         f = AddCellConnToPoints(close_loop=True)
@@ -668,7 +663,7 @@ class TestAddCellConnToPoints(TestBase):
         f.Update()
         output = f.GetOutput()
         # NOTE: the algorithm adds vertice and line cells
-        self.assertEqual(len(self.pts) + len(self.pts), output.GetNumberOfCells())
+        self.assertEqual(len(self.pts), output.GetNumberOfCells())
         # Its fairly difficult to test the nearest neighbor approximations...
         # This was done visually in ParaView.
         # The above test is just there to make sure no errors are thrown
