### 2019-04-25 #1384 @leouieda
Hi @banesullivan, just an update: I'm trying to find reviewers for your two submissions but I'm having a hard time tracking down people with the right balance of skills. I'm still waiting on a few replies but I'm hoping to get this started soon.

I'm also not too familiar with the manuscript building pipeline. Try moving the figure higher up in the document. You can ask Whedon to compile the PDF from a branch so you don't have to make broken commits to *master*: `@whedon generate pdf from branch custom-branch-name`
