### 2019-04-25 #1384 @banesullivan
Thank you for the update, @leouieda!

I'll experiment on another branch with the figure layouts - thanks for the tip!
