+++ _sources/examples/ubc/read-grv-obs.rst.txt
@@ -55,7 +55,7 @@ Download sample data files and keep track of names:
 
  .. code-block:: none
 
-    PolyData (0x7fc03d64c0a8)
+    PolyData (0x7f805aa8cd68)
       N Cells:      191
       N Points:     191
       X Bounds:     3.565e+05, 3.702e+05
@@ -80,7 +80,7 @@ Download sample data files and keep track of names:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.793 seconds)
+   **Total running time of the script:** ( 0 minutes  0.760 seconds)
 
 
 .. _sphx_glr_download_examples_ubc_read-grv-obs.py:
