+++ _sources/examples/filters-general/extract-array.rst.txt
@@ -57,18 +57,18 @@ Construct the filter
 
  .. code-block:: none
 
-    vtkTable (0x4c65a50)
+    vtkTable (0x40fc2b0)
       Debug: Off
-      Modified Time: 21254
+      Modified Time: 21527
       Reference Count: 2
       Registered Events: (none)
-      Information: 0x4343ba0
+      Information: 0x4180380
       Data Released: False
       Global Release Data: Off
-      UpdateTime: 21266
+      UpdateTime: 21539
       Field Data:
         Debug: Off
-        Modified Time: 21248
+        Modified Time: 21521
         Reference Count: 1
         Registered Events: (none)
         Number Of Arrays: 0
@@ -76,7 +76,7 @@ Construct the filter
         Number Of Tuples: 0
       RowData: 
         Debug: Off
-        Modified Time: 21262
+        Modified Time: 21535
         Reference Count: 1
         Registered Events: (none)
         Number Of Arrays: 1
@@ -99,7 +99,7 @@ Construct the filter
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.069 seconds)
+   **Total running time of the script:** ( 0 minutes  0.068 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_extract-array.py:
