### 2019-04-29 0e8a51a
Enable input type preservation LatLon filter
