### 2019-05-01 #1384 @leouieda
:wave: Hi @ngoldbaum @rowanc1 would you be interested in reviewing this submission for JOSS?
