### 2019-05-03 #1384 @leouieda
@gassmoeller thank you agreeing to review :tada: I just wanted to add a disclaimer that @gassmoeller @banesullivan and I have recently been part on the same conference session proposals. However, this does not constitute a conflict of interest [according to our policies](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).
