### 2019-05-08 736c9b0
refactor for changes in vista
