+++ joss/paper.bib
@@ -55,11 +55,11 @@
     journal={ParaView},
 }
 
-@article{vtki,
+@article{vista,
     author = {Sullivan, C. Bane and Kaszynski, Alexander},
     journal = {Journal of Open Source Software},
     pages = {4(36), 1382},
-    title = {{vtki: A Streamlined Python Interface for the Visualization Toolkit}},
+    title = {{vista: A Streamlined Python Interface for the Visualization Toolkit}},
     year = {2019},
     doi = {10.21105/joss.01382},
 }
