+++ PVGeo/filters/tables.py
@@ -69,7 +69,7 @@ class CombineTables(FilterBase):
         self.SetInputDataObject(0, table0)
         self.SetInputDataObject(1, table1)
         self.Update()
-        return interface.wrap_vtki(self.GetOutput())
+        return interface.wrap_vista(self.GetOutput())
 
 
 ###############################################################################
@@ -259,7 +259,7 @@ class ExtractArray(FilterBase):
         arr, field = _helpers.search_for_array(input_data_object, array_name)
         self.SetInputArrayToProcess(0, 0, 0, field, array_name)
         self.Update()
-        return interface.wrap_vtki(self.GetOutput())
+        return interface.wrap_vista(self.GetOutput())
 
 
 
@@ -332,7 +332,7 @@ class SplitTableOnArray(FilterBase):
         arr, field = _helpers.search_for_array(input_data_object, array_name)
         self.SetInputArrayToProcess(0, 0, 0, field, array_name)
         self.Update()
-        return interface.wrap_vtki(self.GetOutput())
+        return interface.wrap_vista(self.GetOutput())
 
 ###############################################################################
 
@@ -401,7 +401,7 @@ class AppendTableToCellData(FilterPreserveTypeBase):
         self.SetInputDataObject(0, dataset)
         self.SetInputDataObject(1, table)
         self.Update()
-        return interface.wrap_vtki(self.GetOutput())
+        return interface.wrap_vista(self.GetOutput())
 
     def get_time_step_values(self):
         """Use this in ParaView decorator to register timesteps.

[…]
