+++ _downloads/3f025dfe6bfa90c3a795e35e98416402/many-slices-along-points.py
@@ -19,14 +19,14 @@ This example demos :class:`PVGeo.filters.ManySlicesAlongPoints`
 """
 ################################################################################
 # sphinx_gallery_thumbnail_number = 3
-import vtki
-from vtki import examples
+import vista
+from vista import examples
 import numpy as np
 import PVGeo
 from PVGeo.filters import ManySlicesAlongPoints
 
 ################################################################################
-# Load a volumetric model to be sliced from vtki's examples
+# Load a volumetric model to be sliced from vista's examples
 model = examples.load_channels()
 model.plot()
 
@@ -43,9 +43,9 @@ def path(y):
 x, y = path(np.arange(model.bounds[2], model.bounds[3], 15.0))
 zo = np.linspace(9.0, 11.0, num=len(y))
 # Make a VTK data object for the filter to use
-points = vtki.PolyData(np.c_[x,y,zo])
+points = vista.PolyData(np.c_[x,y,zo])
 
-p = vtki.Plotter()
+p = vista.Plotter()
 p.add_mesh(model.outline(), color='k')
 p.add_mesh(points, point_size=10.0)
 p.show()
@@ -62,7 +62,7 @@ print(slices)
 
 ################################################################################
 line = PVGeo.filters.AddCellConnToPoints().apply(points)
-p = vtki.Plotter()
+p = vista.Plotter()
 p.add_mesh(line, line_width=10)
 p.add_mesh(slices, lighting=False)
 p.add_mesh(model.outline(), color='k')
