+++ _downloads/463f5044a8b596846a6da59eabd1db80/read-grv-obs.py
@@ -6,8 +6,8 @@ Read a UBC gravity observations file
 """
 # sphinx_gallery_thumbnail_number = 1
 import PVGeo
-import vtki
-from vtki import examples
+import vista
+from vista import examples
 
 ################################################################################
 # Download sample data files and keep track of names:
