+++ _sources/examples/filters-general/normalize-array.rst.txt
@@ -24,8 +24,8 @@ This example demos :class:`PVGeo.filters.NormalizeArray`
 .. code-block:: default
 
     import numpy as np
-    import vtki
-    from vtki import examples
+    import vista
+    from vista import examples
     import PVGeo
     from PVGeo.filters import NormalizeArray
 
@@ -71,7 +71,7 @@ Create some input data. this can be any `vtkDataObject`
 
  .. code-block:: none
 
-    UniformGrid (0x7f3c6ffa46a8)
+    UniformGrid (0x7f78301b9fa8)
       N Cells:      729
       N Points:     1000
       X Bounds:     0.000e+00, 9.000e+00
@@ -96,7 +96,7 @@ Create some input data. this can be any `vtkDataObject`
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  2.233 seconds)
+   **Total running time of the script:** ( 0 minutes  2.220 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_normalize-array.py:
