+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 2.0.0a
+current_version = 2.0.0b
 commit = True
 tag = True
 
