+++ _sources/examples/filters-general/extract-points.rst.txt
@@ -63,7 +63,7 @@ Apply the filter:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.740 seconds)
+   **Total running time of the script:** ( 0 minutes  0.733 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_extract-points.py:
