+++ _sources/examples/filters-general/many-slices-along-points.rst.txt
@@ -112,7 +112,7 @@ slices to what you prefer!
 
  .. code-block:: none
 
-    MultiBlock (0x7f78301b9e88)
+    MultiBlock (0x7f31b302cd68)
       N Blocks:     6
       X Bounds:     0.000, 250.000
       Y Bounds:     0.000, 250.000
@@ -140,7 +140,7 @@ slices to what you prefer!
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  3.664 seconds)
+   **Total running time of the script:** ( 0 minutes  3.665 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_many-slices-along-points.py:
