### 2019-05-12 d6851ad
refactor for PyVista
