+++ joss/paper.bib
@@ -55,11 +55,11 @@
     journal={ParaView},
 }
 
-@article{vista,
+@article{pyvista,
     author = {Sullivan, C. Bane and Kaszynski, Alexander},
     journal = {Journal of Open Source Software},
     pages = {4(36), 1382},
-    title = {{vista: A Streamlined Python Interface for the Visualization Toolkit}},
+    title = {{pyvista: A Streamlined Python Interface for the Visualization Toolkit}},
     year = {2019},
     doi = {10.21105/joss.01382},
 }
