+++ examples/filters-general/points-to-tube.py
@@ -9,7 +9,7 @@ Takes points from a `vtkPolyData` object and constructs a line of those points
 then builds a polygonal tube around that line with some specified radius and number of sides.
 """
 # sphinx_gallery_thumbnail_number = 2
-import vista
+import pyvista
 import numpy as np
 from PVGeo.filters import PointsToTube
 
@@ -23,7 +23,7 @@ def path(y):
 
 x, y = path(np.arange(0.0, 200.0, 25.0))
 zo = np.linspace(9.0, 11.0, num=len(y))
-points = vista.PolyData(np.c_[x,y,zo])
+points = pyvista.PolyData(np.c_[x,y,zo])
 
 points.plot(point_size=10)
 
