+++ PVGeo/grids/fileio.py
@@ -418,7 +418,7 @@ class WriteImageDataToSurfer(WriterBase):
         arr, field = _helpers.search_for_array(input_data_object, array_name)
         self.SetInputArrayToProcess(0, 0, 0, field, array_name)
         self.Update()
-        return interface.wrap_vista(self.GetOutput())
+        return interface.wrap_pyvista(self.GetOutput())
 
     def Write(self, input_data_object=None, array_name=None):
         """Perfrom the write out."""
