+++ _downloads/ec113522a0eb4446c49113b603b55143/read-topo.py
@@ -6,8 +6,8 @@ Read a UBC topography file
 """
 # sphinx_gallery_thumbnail_number = 1
 import PVGeo
-import vista
-from vista import examples
+import pyvista
+from pyvista import examples
 
 ################################################################################
 # Download sample data files and keep track of names:
