+++ _downloads/d808e7ed6a3afd12bcecdc4f1b141ba7/extract-array.py
@@ -7,7 +7,7 @@ to make a `vtkTable` of that single data array
 This example demos :class:`PVGeo.filters.ExtractArray`
 """
 from PVGeo.filters import ExtractArray
-from vista import examples
+from pyvista import examples
 ################################################################################
 # Create input data
 grd = examples.load_rectilinear()
