+++ _downloads/6524aaeafe41cdd93bd5396a139e1378/read-sgems-grid.py
@@ -6,8 +6,8 @@ Read SGeMS Grid file formats. The examples shown here are downloaded from
 `Multiple-point Geostatistics stochastic modeling with training images <http://www.trainingimages.org/training-images-library.html>`_ website.
 """
 # sphinx_gallery_thumbnail_number = 2
-import vista
-from vista import examples
+import pyvista
+from pyvista import examples
 from PVGeo.gslib import SGeMSGridReader
 
 ################################################################################
