+++ _sources/examples/filters-general/voxelize-points.rst.txt
@@ -20,8 +20,8 @@ This example demos :class:`PVGeo.filters.VoxelizePoints`
 .. code-block:: default
 
     # sphinx_gallery_thumbnail_number = 2
-    import vista
-    from vista import examples
+    import pyvista
+    from pyvista import examples
     import numpy as np
     import pandas as pd
     import PVGeo
@@ -53,7 +53,7 @@ but we have forgotten the deatials of the cell spacings and local coordinate
 rotations.
 
 We will read in this data with ``pandas`` and send it to the
-:func:`PVGeo.points_to_poly_data` helper to create a :class:`vista.PolyData`
+:func:`PVGeo.points_to_poly_data` helper to create a :class:`pyvista.PolyData`
 object (essentially a point cloud).
 
 
@@ -94,7 +94,7 @@ object (essentially a point cloud).
 
  .. code-block:: none
 
-    PolyData (0x7f31b3005e88)
+    PolyData (0x7f061775c1c8)
       N Cells:      499200
       N Points:     499200
       X Bounds:     3.268e+05, 3.302e+05
@@ -103,7 +103,7 @@ object (essentially a point cloud).
       N Scalars:    1
 
 
-Note that we have a :class:`vista.PolyData` object now which allows us to do
+Note that we have a :class:`pyvista.PolyData` object now which allows us to do
 all types of immediate plotting of our data. First, lets threshold our points
 as the point cloud has a bunch of zeros and ones throughout the dataspace to
 describe the presence of a fault.
@@ -170,7 +170,7 @@ The full pipeline method
     Voxelizing... done.
     Recovered Angle (deg.): 53.550
     Recovered Cell Sizes: (25.00, 25.00, 25.00)
-    UnstructuredGrid (0x7f31b3005dc8)
+    UnstructuredGrid (0x7f061775c168)
       N Cells:      499200
       N Points:     524064
       X Bounds:     3.268e+05, 3.302e+05
@@ -199,7 +199,7 @@ And now we can plot the voxelized volume
 Filter Volumetric Data
 ++++++++++++++++++++++
 
-Now lets use one of ``vista``'s filters to create slices of the thresholded
+Now lets use one of ``pyvista``'s filters to create slices of the thresholded
 dataset. Specifically, we are using the ``slice_orthogonal`` filter that will
 create 3 orthogonal slices through a data volume.
 
@@ -219,7 +219,7 @@ create 3 orthogonal slices through a data volume.
 
  .. code-block:: none
 
-    MultiBlock (0x7f31b3005d68)
+    MultiBlock (0x7f061775c468)
       N Blocks:     3
       X Bounds:     326804.336, 330185.260
       Y Bounds:     4406253.954, 4409862.606
@@ -245,7 +245,7 @@ Now display the slices and clipped volume
 .. code-block:: default
 
 
-    p = vista.Plotter()
+    p = pyvista.Plotter()
     p.add_mesh(slices)
     p.add_mesh(clip)
     p.show_grid()
@@ -262,7 +262,7 @@ Now display the slices and clipped volume
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  27.691 seconds)
+   **Total running time of the script:** ( 0 minutes  28.844 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_voxelize-points.py:
