+++ _sources/examples/filters-general/extract-points.rst.txt
@@ -18,7 +18,7 @@ This example demos :class:`PVGeo.filters.ExtractPoints`
 .. code-block:: default
 
     # sphinx_gallery_thumbnail_number = 2
-    from vista import examples
+    from pyvista import examples
     from PVGeo.filters import ExtractPoints
 
 
@@ -63,7 +63,7 @@ Apply the filter:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.733 seconds)
+   **Total running time of the script:** ( 0 minutes  0.762 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_extract-points.py:
