+++ setup.py
@@ -26,7 +26,7 @@ install_requires=[
     'pandas>=0.23.4',
     'mock>=2.0.0',
     'espatools>=0.0.8',
-    'pyvista>=0.20.0'
+    'pyvista>=0.20.1'
 ]
 
 # add vtk if not windows and (not Python 3.x or not x64)
