### 2019-05-13 cfe2ca3
More pep8 refactoring
