### 2019-05-13 a185f06
More pep8 refactoring
