+++ PVGeo/grids/fileio.py
@@ -263,13 +263,13 @@ class SurferGridReader(ReaderBase):
         return grd
 
 
-    def _ReadGrids(self, idx=None):
+    def _read_grids(self, idx=None):
         """This parses the first file to determine grid file type then reads
         all files set."""
         if idx is not None:
-            filenames = [self.GetFileNames(idx=idx)]
+            filenames = [self.get_file_names(idx=idx)]
         else:
-            filenames = self.GetFileNames()
+            filenames = self.get_file_names()
         contents = []
         f = open(filenames[0], 'rb')
         key = struct.unpack('4s', f.read(4))[0]
@@ -297,7 +297,7 @@ class SurferGridReader(ReaderBase):
         """Should not need to be overridden.
         """
         # Perform Read
-        self.__grids = self._ReadGrids()
+        self.__grids = self._read_grids()
         self.need_to_read(flag=False)
         return 1
 
@@ -340,7 +340,7 @@ class SurferGridReader(ReaderBase):
             self.__data_name = data_name
             self.Modified(read_again=False)
 
-    def GetDataName(self):
+    def get_data_name(self):
         """Get the name of the data array"""
         return self.__data_name
 
@@ -431,6 +431,10 @@ class WriteImageDataToSurfer(WriterBase):
         self.Update()
 
 
+    def write(self, input_data_object=None, array_name=None):
+        """Perfrom the write out."""
+        return self.Write(input_data_object=input_data_object, array_name=array_name)
+
 ###############################################################################
 
 
@@ -453,7 +457,7 @@ class EsriGridReader(DelimitedTextReader):
         self.__data_name = kwargs.get('data_name', 'Data')
         self.NODATA_VALUE = -9999
 
-    def _ExtractHeader(self, content):
+    def _extract_header(self, content):
         """Internal helper to parse header information in ESRI Grid files"""
         try:
             self.__nx = int(content[0].split()[1])
@@ -466,13 +470,13 @@ class EsriGridReader(DelimitedTextReader):
             raise _helpers.PVGeoError('This file is not in proper Esri ASCII Grid format.')
         return [self.__data_name], content[6::]
 
-    def _FileContentsToDataFrame(self, contents):
+    def _file_contents_to_data_frame(self, contents):
         """Creates a dataframe with a sinlge array for the file data.
         """
         data = []
         for content in contents:
             arr = np.fromiter((float(s) for line in content for s in line.split()), dtype=float)
-            df = pd.DataFrame(data=arr, columns=[self.GetDataName()])
+            df = pd.DataFrame(data=arr, columns=[self.get_data_name()])
             data.append(df)
         return data
 
@@ -535,7 +539,7 @@ class EsriGridReader(DelimitedTextReader):
             self.__data_name = data_name
             self.Modified(read_again=False)
 
-    def GetDataName(self):
+    def get_data_name(self):
         """Get the name of the data array"""
         return self.__data_name
 
@@ -569,21 +573,24 @@ class LandsatReader(ReaderBaseBase):
         """
         return ReaderBaseBase.Modified(self, read_again=read_again)
 
+    def modified(self, read_again=False):
+        return self.Modified(read_again=read_again)
 
-    def GetFileName(self):
+
+    def get_file_name(self):
         """Super class has file names as a list but we will only handle a single
         project file. This provides a conveinant way of making sure we only
         access that single file.
-        A user could still access the list of file names using ``GetFileNames()``.
+        A user could still access the list of file names using ``get_file_names()``.
         """
-        return ReaderBaseBase.GetFileNames(self, idx=0)
[…]
