+++ _sources/examples/filters-general/append-cell-centers.rst.txt
@@ -44,7 +44,7 @@ Use an example mesh from pyvista
 
  .. code-block:: none
 
-    RectilinearGrid (0x7f0617806d08)
+    RectilinearGrid (0x7f83b001bbe8)
       N Cells:      16146
       N Points:     18144
       X Bounds:     -3.500e+02, 1.350e+03
@@ -72,7 +72,7 @@ Run the PVGeo algorithm
 
  .. code-block:: none
 
-    RectilinearGrid (0x7f0617806e28)
+    RectilinearGrid (0x7f83b001bd08)
       N Cells:      16146
       N Points:     18144
       X Bounds:     -3.500e+02, 1.350e+03
@@ -98,7 +98,7 @@ Run the PVGeo algorithm
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.812 seconds)
+   **Total running time of the script:** ( 0 minutes  0.807 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_append-cell-centers.py:
