### 2019-05-13 5cb36e4
Bump version: 2.0.0b → 2.0.0
