+++ PVGeo/__init__.py
@@ -46,7 +46,7 @@ from . import version
 __author__ = 'Bane Sullivan'
 __license__ = 'BSD-3-Clause'
 __copyright__ = '2018, Bane Sullivan'
-__version__ = '2.0.0b'
+__version__ = '2.0.0'
 __displayname__ = 'PVGeo'
 
 
