+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 2.0.0b
+current_version = 2.0.0
 commit = True
 tag = True
 
