+++ _sources/examples/gslib/sg_execution_times.rst.txt
@@ -5,7 +5,7 @@
 
 Computation times
 =================
-**00:48.994** total execution time for **examples_gslib** files:
+**00:49.039** total execution time for **examples_gslib** files:
 
-- **00:48.491**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
-- **00:00.503**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
+- **00:48.547**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
+- **00:00.492**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
