### 2019-05-13 b1780d6
Update PyVista descriptions
