+++ _downloads/4f47c906fbfeea4520d436b6d55f084b/read-tensor-mesh.py
@@ -25,7 +25,7 @@ mesh = reader.apply()
 print(mesh)
 
 ################################################################################
-# Use a ``pyvista`` ``threshold`` filter to remove ``NaN`` data values
+# Use a `PyVista` ``threshold`` filter to remove ``NaN`` data values
 mesh.threshold().plot()
 
 
