+++ _sources/examples/ubc/read-tensor-mesh.rst.txt
@@ -62,7 +62,7 @@ Read the mesh and model
 
  .. code-block:: none
 
-    RectilinearGrid (0x7fccb6f71228)
+    RectilinearGrid (0x7f002e1721c8)
       N Cells:      190440
       N Points:     200900
       X Bounds:     3.550e+05, 3.722e+05
@@ -72,7 +72,7 @@ Read the mesh and model
       N Scalars:    1
 
 
-Use a ``pyvista`` ``threshold`` filter to remove ``NaN`` data values
+Use a `PyVista` ``threshold`` filter to remove ``NaN`` data values
 
 
 .. code-block:: default
@@ -124,7 +124,7 @@ Or threshold a data range
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  4.346 seconds)
+   **Total running time of the script:** ( 0 minutes  4.028 seconds)
 
 
 .. _sphx_glr_download_examples_ubc_read-tensor-mesh.py:
