+++ joss/paper.bib
@@ -59,7 +59,7 @@
     author = {Sullivan, C. Bane and Kaszynski, Alexander},
     journal = {Journal of Open Source Software},
     pages = {4(36), 1382},
-    title = {{pyvista: A Streamlined Python Interface for the Visualization Toolkit}},
+    title = {{PyVista: A Streamlined Python Interface for the Visualization Toolkit}},
     year = {2019},
     doi = {10.21105/joss.01382},
 }
