+++ _sources/examples/filters-general/array-math.rst.txt
@@ -72,7 +72,7 @@ Use the filter:
 
  .. code-block:: none
 
-    UniformGrid (0x7f002e1ad588)
+    UniformGrid (0x7f35021432e8)
       N Cells:      243
       N Points:     400
       X Bounds:     0.000e+00, 9.000e+00
@@ -122,7 +122,7 @@ Use a custom math operation:
 
  .. code-block:: none
 
-    UniformGrid (0x7f002e1ad888)
+    UniformGrid (0x7f35021435e8)
       N Cells:      243
       N Points:     400
       X Bounds:     0.000e+00, 9.000e+00
@@ -147,7 +147,7 @@ Use a custom math operation:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.067 seconds)
+   **Total running time of the script:** ( 0 minutes  0.065 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_array-math.py:
