+++ _sources/examples/gslib/read-point-set.rst.txt
@@ -44,7 +44,7 @@ Read GSLib point set file
 
  .. code-block:: none
 
-    PolyData (0x7f002e172348)
+    PolyData (0x7f350211ff48)
       N Cells:      100
       N Points:     100
       X Bounds:     3.000e+00, 2.520e+02
@@ -69,7 +69,7 @@ Read GSLib point set file
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.495 seconds)
+   **Total running time of the script:** ( 0 minutes  0.471 seconds)
 
 
 .. _sphx_glr_download_examples_gslib_read-point-set.py:
