### 2019-05-15 #1384 @banesullivan
@leouieda All is stable and PVGeo is released on version `2.0.0` which is the version I'd like to submit for review (not `1.2.3` mentioned above)
