+++ _sources/examples/gslib/sg_execution_times.rst.txt
@@ -5,7 +5,7 @@
 
 Computation times
 =================
-**00:48.373** total execution time for **examples_gslib** files:
+**00:48.724** total execution time for **examples_gslib** files:
 
-- **00:47.902**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
-- **00:00.471**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
+- **00:48.144**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
+- **00:00.581**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
