### 2019-05-15 #1451 @whedon
[REVIEW]: PVGeo: an open-source Python package for geoscientific visualization in VTK and ParaView
**Submitting author:** @banesullivan (<a href="http://orcid.org/0000-0001-8628-4566">C. Bane Sullivan</a>)
**Repository:** <a href="https://github.com/OpenGeoVis/PVGeo" target ="_blank">https://github.com/OpenGeoVis/PVGeo</a>
**Version:** v2.0.1
