### 2019-05-16 #1384 @alexgarciac
hi. as this is python and a py package there should be some sample data so that anyone could test that the package and all the assertions being made in the paper do really work. I read the paper and there is a lot about a lot but there is no sample data to test the package.
