### 2019-05-16 b65c352
Update dependencies in docs
