### 2019-05-16 30e09ce
Auto deploy example notebooks to binder repo
