+++ docs/deploy_docs_from_travis.sh
@@ -1,0 +1,19 @@
+#!/bin/sh
+# MAke sure this is called from root level of repository
+set -x
+cookiecutter -f --no-input --config-file ./docs/pvgeo-binder-config.yml -o .. https://github.com/pyvista/cookiecutter-pyvista-binder.git;
+rm -rf ../pvgeo-binder/notebooks/;
+cd ./docs/;
+find ./examples -type f -name '*.ipynb' | cpio -p -d -v ../../pvgeo-binder/;
+cd ../../pvgeo-binder/;
+git init;
+git add .;
+git commit -m "${TRAVIS_JOB_NUMBER} : Autogenerated notebooks from Travis";
+REMOTE="https://${GH_TOKEN}@github.com/OpenGeoVis/pvgeo-binder";
+git config --global user.name "${GH_NAME}";
+git config --global user.email "${GH_EMAIL}";
+git remote add origin ${REMOTE};
+git push -uf origin master;
+cd ../PVGeo/
+doctr deploy --built-docs ./docs/_build/html .;
+set +x
