+++ .travis.yml
@@ -34,16 +34,6 @@ script:
     - coverage run tests/__main__.py
     - codecov --token=8da9d332-1618-450d-8654-09a8ecf3c2dc
     - python tests/__main__.py close # exit with error code if problem occurs
-    # Make and deploy docs - I have no idea why but it always segfault on first make - so make twice
-    - if [[($TRAVIS_PYTHON_VERSION == 3.7)]]; then
-        echo "Building and uploading the docs";
-        cd ./docs;
-        make html;
-        set -e;
-        make html;
-        cd ..;
-        doctr deploy --built-docs ./docs/build/html/ .;
-      fi
 
 notifications:
     slack:
@@ -58,7 +48,12 @@ notifications:
         on_failure: never
 
 deploy:
-    provider: pypi
+  - provider: script
+    script: sh ./docs/deploy_docs_from_travis.sh
+    skip_cleanup: true
+    on:
+      condition: ($TRAVIS_PYTHON_VERSION == 3.7) && ($TRAVIS_REPO_SLUG == 'OpenGeoVis/PVGeo') && ($TRAVIS_BRANCH == 'master') && ($DOCS_TRIGGER == 'BANE_RULES')
+  - provider: pypi
     user: "pvgeosupport"
     password:
         secure: L3eSpvBG9sJeGpUCybkisOdpjIv3BxEy2BPHDimSrJE26UnJ4my/8Xq05GWejH4IOBhyo6ZmlKy2olZeTaIY+0+kf2zpR8C0b5DW4B2mXKkiEm7JpVdWmlc/Kcw7aBOvn3APH5K52if58p3M65P+9nkKE91FdmZUQP41e32lHl6yYMo8PBU8x7imMdL4Myl5J+5DxtYnr4tC07Td4KTvhNJyUM0Y5i3Qyp1AELlB36M6gcvUnRCAKqQDq5phtufrQHM/TjAeyuX2RwzDtZhILWEYHYPDZ0m8emPV5OYbklqHaEM5DcY4Vd50l/yw22mm154nqsOUr3nz2URLV5u3qea+5WZeQRaDjFNVHe0udO8axhT8I2VPyatnCX2q9GmDCF77Dl9Fv/ZLXP80qanL1BfwjPIB1vkMPR+6y0n8DcI3MM5SMZYm80igPLY/AMqBFsWoU9Fsp+luvx6Dcmf3s7RTKf0tqOJcog07GRzK0c4up8uZwBir3rjQzhv4y0DT3IfillQB7SmMzH4xW0vpi/txvaBNvlI0OT3eIiygH79FoxzWZO0YdzZLV33ZUU7nKk+TsQka4LmjksVQw2+aK5fVoJ9VuUVYYfZDtoJICKfktfv7UFdl0tJkjHArSOrPVmsrit54uESmKsMIMzeQi+fY/Baz63utkXFm+8U6D3k=
