### 2019-05-16 b6dc5a4
Fix doctr deploy script
