### 2019-05-16 #1384 @leouieda
Hi @alexgarciac, the JOSS paper [should not contain documentation or data](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). These are expect to be included in the software documentation, which the submission does. For example: https://pvgeo.org/examples/filters-general/voxelize-points.html
