### 2019-05-16 d95dee5
Update binder deploy script
