+++ docs/deploy_docs_from_travis.sh
@@ -3,9 +3,12 @@
 set -x
 cookiecutter -f --no-input --config-file ./docs/pvgeo-binder-config.yml -o .. https://github.com/pyvista/cookiecutter-pyvista-binder.git;
 rm -rf ../pvgeo-binder/notebooks/;
+rm -rf ../pvgeo-binder/examples/;
 cd ./docs/;
-find ./examples -type f -name '*.ipynb' | cpio -p -d -v ../../pvgeo-binder/;
+find ./source/examples -type f -name '*.ipynb' | cpio -p -d -v ../../pvgeo-binder/;
 cd ../../pvgeo-binder/;
+mv ./source/examples/ ./examples/
+rm -r ./source/
 git init;
 git add .;
 git commit -m "${TRAVIS_JOB_NUMBER} : Autogenerated notebooks from Travis";
