+++ _sources/examples/grids/read-surfer.rst.txt
@@ -59,7 +59,7 @@ Apply a filter to the DEM to have realistic topography
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  1.299 seconds)
+   **Total running time of the script:** ( 0 minutes  1.379 seconds)
 
 
 .. _sphx_glr_download_examples_grids_read-surfer.py:
