### 2019-05-16 8f6f479
Install cookiecutter on Travis
