+++ .travis.yml
@@ -27,6 +27,7 @@ install:
     - pip install -r requirements.txt
     - if [[($TRAVIS_PYTHON_VERSION == 3.7)]]; then
         pip install -r ./docs/requirements.txt;
+        pip install cookiecutter;
       fi
     - python setup.py install
 
