### 2019-05-16 c9624df
Fix binder examples deploy
