+++ docs/pvgeo-binder-config.yml
@@ -1,5 +1,5 @@
 default_context:
-    repo_name: "PVGeo-examples"
+    repo_name: "pvgeo-binder"
     github_account: "OpenGeoVis"
     description: "Quick examples demonstrating what PVGeo can do! Simply launch this Binder project and test drive PVGeo"
     author_name: "PVGeo Developers"
