### 2019-05-16 d75132a
Fix requirements for binder examples
