+++ docs/deploy_docs_from_travis.sh
@@ -9,6 +9,10 @@ find ./source/examples -type f -name '*.ipynb' | cpio -p -d -v ../../pvgeo-binde
 cd ../../pvgeo-binder/;
 mv ./source/examples/ ./examples/
 rm -r ./source/
+# Add requirements to environment.yml
+echo "    - git+git://github.com/OpenGeoVis/PVGeo@master" >> environment.yml
+echo "    - omfvista>=0.2.0" >> environment.yml
+echo "    - discretize>=0.4.3" >> environment.yml
 git init;
 git add .;
 git commit -m "${TRAVIS_JOB_NUMBER} : Autogenerated notebooks from Travis";
