+++ requirements.txt
@@ -10,4 +10,4 @@ pyproj>=1.9
 omf>=0.9.3
 omfvista>=0.2.0
 espatools>=0.0.8
-discretize>=0.3.8
+discretize>=0.4.3
