+++ joss/paper.bib
@@ -17,10 +17,11 @@
 
 
 @book{vtkbook,
-    title={The visualization toolkit: an object-oriented approach to 3D graphics},
+    title={The Visualization Toolkit (4th ed.)},
     author={Schroeder, Will J and Lorensen, Bill and Martin, Ken},
-    year={2004},
-    publisher={Kitware}
+    year={2006},
+    publisher={Kitware},
+    isbn={978-1-930934-19-1}
 }
 
 
@@ -58,8 +59,8 @@
 @article{pyvista,
     author = {Sullivan, C. Bane and Kaszynski, Alexander},
     journal = {Journal of Open Source Software},
-    pages = {4(36), 1382},
-    title = {{PyVista: A Streamlined Python Interface for the Visualization Toolkit}},
+    pages = {4(37), 1450},
+    title = {{PyVista: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK)}},
     year = {2019},
-    doi = {10.21105/joss.01382},
+    doi = {10.21105/joss.01450},
 }
