### 2019-05-18 a530a04
Minor updates to JOSS paper
