+++ _sources/examples/ubc/read-tensor-mesh.rst.txt
@@ -62,7 +62,7 @@ Read the mesh and model
 
  .. code-block:: none
 
-    RectilinearGrid (0x7fa39432a228)
+    RectilinearGrid (0x7f5d3da4f768)
       N Cells:      190440
       N Points:     200900
       X Bounds:     3.550e+05, 3.722e+05
@@ -124,7 +124,7 @@ Or threshold a data range
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  4.285 seconds)
+   **Total running time of the script:** ( 0 minutes  4.274 seconds)
 
 
 .. _sphx_glr_download_examples_ubc_read-tensor-mesh.py:
