### 2019-05-18 a18be99
Fix typo in JOSS paper
