### 2019-05-18 dc936a1
Remove FORGE figure from JOSS paper
