+++ _sources/examples/gslib/sg_execution_times.rst.txt
@@ -5,7 +5,7 @@
 
 Computation times
 =================
-**00:49.279** total execution time for **examples_gslib** files:
+**00:48.940** total execution time for **examples_gslib** files:
 
-- **00:48.800**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
-- **00:00.479**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
+- **00:48.451**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
+- **00:00.488**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
