### 2019-05-22 870a7cc
Update PyVista citation
