+++ joss/paper.bib
@@ -56,11 +56,16 @@
     journal={ParaView},
 }
 
-@article{pyvista,
-    author = {Sullivan, C. Bane and Kaszynski, Alexander},
-    journal = {Journal of Open Source Software},
-    pages = {4(37), 1450},
-    title = {{PyVista: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK)}},
-    year = {2019},
+@article{sullivan2019pyvista,
     doi = {10.21105/joss.01450},
+    url = {https://doi.org/10.21105/joss.01450},
+    year = {2019},
+    month = {may},
+    publisher = {The Open Journal},
+    volume = {4},
+    number = {37},
+    pages = {1450},
+    author = {C. Bane Sullivan and Alexander Kaszynski},
+    title = {{PyVista}: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit ({VTK})},
+    journal = {Journal of Open Source Software}
 }
