### 2019-05-22 f6828b6
Update docs after building Travis build 813 of OpenGeoVis/PVGeo
The docs were built from the branch 'master' against the commit
870a7cc0751cc45a31b498e0ed46f98e4303517f.

The Travis build that generated this commit is at
https://travis-ci.org/OpenGeoVis/PVGeo/jobs/535597458.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy --built-docs ./docs/build/html .
