+++ _sources/examples/filters-general/normalize-array.rst.txt
@@ -71,7 +71,7 @@ Create some input data. this can be any `vtkDataObject`
 
  .. code-block:: none
 
-    UniformGrid (0x7f76542f9fa8)
+    UniformGrid (0x7f23c252b8e8)
       N Cells:      729
       N Points:     1000
       X Bounds:     0.000e+00, 9.000e+00
@@ -98,7 +98,7 @@ Create some input data. this can be any `vtkDataObject`
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  2.268 seconds)
+   **Total running time of the script:** ( 0 minutes  2.292 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_normalize-array.py:
