### 2019-06-05 #1451 @Leguark
Hi @leouieda and @banesullivan

sorry for the delay. First of all, thank you for the authors for the great work on this library. After couple of years dealing myself with vtk what you have done here is going to facilitate enormously how visualize geodata in 3D.

- pip installation works without further issues
- All test passed with a satisfactory coverage
- Most of the methods at least a short explanation of the functionality. Ideally, the most exposed methods--i.e. those that a user must called--would benefit of a longer description of parameters, attributes and returns.
- All examples are working and very well documented which makes up for the lack of more complete docstrings.

In short, the library has already solid foundations to keep building functionality on top. I did not find any of the joos criteria missing. Therefore, I recommend the publication of this library and I am looking forward the future developments of this library and its integration with the rest of open geoscientific tools on the following months.
