### 2019-06-07 OpenGeoVis/PVGeo#49 @gassmoeller
Maybe mention the contribution guide in README.md
Many github projects have started to follow Github's suggestion of having a `CONTRIBUTING.md` file in the main folder to explain to newcomers how to contribute. You do not really need that, as you already have nice instructions at https://pvgeo.org/dev-guide/contributing.html (and in https://github.com/OpenGeoVis/PVGeo/blob/master/docs/source/dev-guide/contributing.rst), but it would be helpful if that were linked in the README.md so that users do not have to search for it (if they find PVGeo through github before finding the website).
