### 2019-06-07 OpenGeoVis/PVGeo#49 @gassmoeller
related to openjournals/joss-reviews#1451
