### 2019-06-08 #1451 @gassmoeller
Dear @leouieda and @banesullivan

I have now finished my review of PVGeo and the accompanying publication in JOSS and can fully recommend its publication. I have verified that the software works as described and I am convinced that it addresses an important need for Geoscientists that can be well extended to more applications in the future. The repository structure is logical and follows best practices and the documentation allows users to understand its purpose and integrate it into their own workflow. The software follows best practices regarding testing, coding, and distribution. The JOSS manuscript clearly describes the need for and application of the software and references relevant projects and publications.

I have opened a minor issue that I would like to see addressed to make it simpler for new users to contribute to the development (see OpenGeoVis/PVGeo#49), and this review issue should be updated to reflect the authors request to submit version 2.0.0 instead of 1.2.3, but otherwise I recommend this paper for timely publication.

Sincerely,
Rene Gassmoeller
