### 2019-06-08 #1451 @banesullivan
Hi @Leguark and @gassmoeller, thank you for the speedy reviews! I have been traveling lately and went offline for about a week so I'm happy to address any issues now.

> I have opened a minor issue that I would like to see addressed to make it simpler for new users to contribute to the development (see OpenGeoVis/PVGeo#49)

Thanks for the suggestion, @gassmoeller. I just updated PVGeo's readme to address this.

> this review issue should be updated to reflect the authors request to submit version 2.0.0 instead of 1.2.3

Ah yes, version 2.0.0 (actually 2.0.1 for when I tag and archive the repo for the submission)
