+++ _sources/examples/filters-general/append-cell-centers.rst.txt
@@ -44,7 +44,7 @@ Use an example mesh from pyvista
 
  .. code-block:: none
 
-    RectilinearGrid (0x7f23c3cd72e8)
+    RectilinearGrid (0x7fb3aa9844c8)
       N Cells:      16146
       N Points:     18144
       X Bounds:     -3.500e+02, 1.350e+03
@@ -72,7 +72,7 @@ Run the PVGeo algorithm
 
  .. code-block:: none
 
-    RectilinearGrid (0x7f23c3cd7408)
+    RectilinearGrid (0x7fb3aa9845e8)
       N Cells:      16146
       N Points:     18144
       X Bounds:     -3.500e+02, 1.350e+03
@@ -98,7 +98,7 @@ Run the PVGeo algorithm
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.825 seconds)
+   **Total running time of the script:** ( 0 minutes  0.815 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_append-cell-centers.py:
