+++ _sources/examples/filters-general/combine-tables.rst.txt
@@ -76,7 +76,7 @@ Create some input tables
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.069 seconds)
+   **Total running time of the script:** ( 0 minutes  0.065 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_combine-tables.py:
