### 2019-06-26 #1451 @leouieda
@banesullivan sorry, it did fall off the radar for a while. Thanks for the reminder :slightly_smiling_face:

@gassmoeller and @Leguark thank you for the reviews! Regarding the version number, it's OK that there is a newer release. We'll update the version number here when it's time for publication. I see that you two **still have unchecked items** in your checklists. Is there anything that still needs to be addressed for those items? If not, please check them so we can move forward.
