### 2019-06-26 50063f1
Minor edits to the JOSS paper
Went through the paper once more and caught a few typos and repetitions.
