+++ _sources/examples/filters-general/voxelize-points.rst.txt
@@ -17,6 +17,7 @@ volumetric operations.
 
 This example demos :class:`PVGeo.filters.VoxelizePoints`
 
+
 .. code-block:: default
 
     # sphinx_gallery_thumbnail_number = 2
@@ -72,9 +73,10 @@ object (essentially a point cloud).
 
  .. code-block:: none
 
-    X            Y       Z  Fault
+                X            Y       Z  Fault
     0  326819.497  4407450.636  1287.5      0
     1  326834.340  4407470.753  1287.5      0
+
 
 
 
@@ -94,13 +96,15 @@ object (essentially a point cloud).
 
  .. code-block:: none
 
-    PolyData (0x7fb3aa900a08)
+    PolyData (0x7f5bf67739a8)
       N Cells:      499200
       N Points:     499200
       X Bounds:     3.268e+05, 3.302e+05
       Y Bounds:     4.406e+06, 4.410e+06
       Z Bounds:     1.250e+01, 1.288e+03
-      N Scalars:    1
+      N Arrays:     1
+
+
 
 
 Note that we have a :class:`pyvista.PolyData` object now which allows us to do
@@ -170,13 +174,15 @@ The full pipeline method
     Voxelizing... done.
     Recovered Angle (deg.): 53.550
     Recovered Cell Sizes: (25.00, 25.00, 25.00)
-    UnstructuredGrid (0x7fb3aa900948)
+    UnstructuredGrid (0x7f5bf6773108)
       N Cells:      499200
       N Points:     524064
       X Bounds:     3.268e+05, 3.302e+05
       Y Bounds:     4.406e+06, 4.410e+06
       Z Bounds:     0.000e+00, 1.300e+03
-      N Scalars:    3
+      N Arrays:     3
+
+
 
 
 And now we can plot the voxelized volume
@@ -219,11 +225,13 @@ create 3 orthogonal slices through a data volume.
 
  .. code-block:: none
 
-    MultiBlock (0x7fb3aa900c48)
+    MultiBlock (0x7f5bf6773b88)
       N Blocks:     3
       X Bounds:     326804.336, 330185.260
       Y Bounds:     4406253.954, 4409862.606
       Z Bounds:     0.000, 1300.000
+
+
 
 
 And let's use a ``clip`` filter:
@@ -262,7 +270,7 @@ Now display the slices and clipped volume
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  29.163 seconds)
+   **Total running time of the script:** ( 0 minutes  29.217 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_voxelize-points.py:
@@ -290,4 +298,4 @@ Now display the slices and clipped volume
 
  .. rst-class:: sphx-glr-signature
 
-    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_
