+++ _sources/examples/filters-general/reshape-table.rst.txt
@@ -16,6 +16,7 @@ This filter will take a `vtkTable` object and reshape it. This filter essentiall
 
 This example demos :class:`PVGeo.filters.ReshapeTable`
 
+
 .. code-block:: default
 
     import numpy as np
@@ -98,7 +99,7 @@ Check the output against NumPy
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.067 seconds)
+   **Total running time of the script:** ( 0 minutes  0.066 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_reshape-table.py:
@@ -126,4 +127,4 @@ Check the output against NumPy
 
  .. rst-class:: sphx-glr-signature
 
-    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_
