+++ _sources/examples/filters-general/combine-tables.rst.txt
@@ -15,6 +15,7 @@ same number of rows into a single `vtkTable`.
 
 This example demos :class:`PVGeo.filters.CombineTables`
 
+
 .. code-block:: default
 
     import numpy as np
@@ -104,4 +105,4 @@ Create some input tables
 
  .. rst-class:: sphx-glr-signature
 
-    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_
