+++ _sources/examples/filters-general/many-slices-along-points.rst.txt
@@ -23,6 +23,7 @@ ParaView to slice that entire model along the travel path to have a single
 warped slice.
 
 This example demos :class:`PVGeo.filters.ManySlicesAlongPoints`
+
 sphinx_gallery_thumbnail_number = 3
 
 
@@ -112,11 +113,13 @@ slices to what you prefer!
 
  .. code-block:: none
 
-    MultiBlock (0x7fb3aa9338e8)
+    MultiBlock (0x7f5bf67a6828)
       N Blocks:     6
       X Bounds:     0.000, 250.000
       Y Bounds:     0.000, 250.000
       Z Bounds:     0.000, 100.000
+
+
 
 
 
@@ -140,7 +143,7 @@ slices to what you prefer!
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  3.737 seconds)
+   **Total running time of the script:** ( 0 minutes  3.783 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_many-slices-along-points.py:
@@ -168,4 +171,4 @@ slices to what you prefer!
 
  .. rst-class:: sphx-glr-signature
 
-    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_
