+++ _sources/examples/gslib/read-point-set.rst.txt
@@ -12,6 +12,7 @@ Read GSLib Point Set
 
 Read GSLib point set file
 
+
 .. code-block:: default
 
     # sphinx_gallery_thumbnail_number = 1
@@ -44,13 +45,15 @@ Read GSLib point set file
 
  .. code-block:: none
 
-    PolyData (0x7fb3aa942ca8)
+    PolyData (0x7f5bf7715948)
       N Cells:      100
       N Points:     100
       X Bounds:     3.000e+00, 2.520e+02
       Y Bounds:     1.000e+00, 2.980e+02
       Z Bounds:     0.000e+00, 0.000e+00
-      N Scalars:    1
+      N Arrays:     1
+
+
 
 
 
@@ -69,7 +72,7 @@ Read GSLib point set file
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.516 seconds)
+   **Total running time of the script:** ( 0 minutes  0.484 seconds)
 
 
 .. _sphx_glr_download_examples_gslib_read-point-set.py:
@@ -97,4 +100,4 @@ Read GSLib point set file
 
  .. rst-class:: sphx-glr-signature
 
-    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_
