+++ _sources/examples/ubc/read-topo.rst.txt
@@ -12,6 +12,7 @@ Read UBC Topography File
 
 Read a UBC topography file
 
+
 .. code-block:: default
 
     # sphinx_gallery_thumbnail_number = 1
@@ -55,13 +56,15 @@ Download sample data files and keep track of names:
 
  .. code-block:: none
 
-    PolyData (0x7fb3aa8a1348)
+    PolyData (0x7f5bd0b63288)
       N Cells:      381924
       N Points:     381924
       X Bounds:     3.550e+05, 3.720e+05
       Y Bounds:     5.999e+06, 6.016e+06
       Z Bounds:     2.050e+03, 3.104e+03
-      N Scalars:    1
+      N Arrays:     1
+
+
 
 
 
@@ -80,7 +83,7 @@ Download sample data files and keep track of names:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  4.370 seconds)
+   **Total running time of the script:** ( 0 minutes  4.314 seconds)
 
 
 .. _sphx_glr_download_examples_ubc_read-topo.py:
@@ -108,4 +111,4 @@ Download sample data files and keep track of names:
 
  .. rst-class:: sphx-glr-signature
 
-    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_
