+++ _sources/examples/filters-general/points-to-tube.rst.txt
@@ -16,6 +16,7 @@ This example will demonstrate how to to build a tube from a set of points in
 Takes points from a `vtkPolyData` object and constructs a line of those points
 then builds a polygonal tube around that line with some specified radius and number of sides.
 
+
 .. code-block:: default
 
     # sphinx_gallery_thumbnail_number = 2
@@ -73,7 +74,7 @@ then builds a polygonal tube around that line with some specified radius and num
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.420 seconds)
+   **Total running time of the script:** ( 0 minutes  0.429 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_points-to-tube.py:
@@ -101,4 +102,4 @@ then builds a polygonal tube around that line with some specified radius and num
 
  .. rst-class:: sphx-glr-signature
 
-    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_

[…]
