### 2019-06-27 #1451 @leouieda
@Leguark seems like Whedon un-assigned you temporarily from the issue for some reason. I've gone ahead and checked your boxes.

@gassmoeller thanks for the confirmation!
