### 2019-06-27 c6fdc37
Remove FORGE figure from JOSS paper (#48)
* Move FORGE figure in JOSS paper

* Reduce size of FORGE image for JOSS paper

* Remove FORGE figure from JOSS paper
