### 2019-06-27 1ac0e50
Fix citation for PyVista
The key in the bibtex file didn't match the citation in the paper.

openjournals/joss-reviews#1451
