### 2019-06-27 #1451 @leouieda
@banesullivan since both reviewers are satisfied with this submission I'm pleased to move ahead with acceptance :tada:

Here are the steps you need to take now:

- Do a final proof-reading of the paper and fix any outstanding issues (please link them to this issue)
- Make a release with the latest changes from the review and post the version number here.
- Archive the release on Zenodo/figshare and post the DOI here.
