### 2019-06-27 bcc0106
Fix citation for PyVista (#51)
The key in the bibtex file didn't match the citation in the paper.

openjournals/joss-reviews#1451
