+++ _sources/examples/filters-general/reshape-table.rst.txt
@@ -99,7 +99,7 @@ Check the output against NumPy
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.066 seconds)
+   **Total running time of the script:** ( 0 minutes  0.088 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_reshape-table.py:
