### 2019-06-27 aa2af9d
Update docs after building Travis build 819 of OpenGeoVis/PVGeo
The docs were built from the branch 'master' against the commit
bcc0106fbb54c46c7ea6034103a12f2f4f079487.

The Travis build that generated this commit is at
https://travis-ci.org/OpenGeoVis/PVGeo/jobs/551087970.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy --built-docs ./docs/build/html .
