+++ _sources/examples/filters-general/array-math.rst.txt
@@ -73,7 +73,7 @@ Use the filter:
 
  .. code-block:: none
 
-    UniformGrid (0x7efc9e684588)
+    UniformGrid (0x7f5bac9e0888)
       N Cells:      243
       N Points:     400
       X Bounds:     0.000e+00, 9.000e+00
@@ -125,7 +125,7 @@ Use a custom math operation:
 
  .. code-block:: none
 
-    UniformGrid (0x7efc9e684a68)
+    UniformGrid (0x7f5bac9e0d68)
       N Cells:      243
       N Points:     400
       X Bounds:     0.000e+00, 9.000e+00
@@ -152,7 +152,7 @@ Use a custom math operation:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.090 seconds)
+   **Total running time of the script:** ( 0 minutes  0.066 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_array-math.py:
