+++ _sources/examples/ubc/read-grv-obs.rst.txt
@@ -56,7 +56,7 @@ Download sample data files and keep track of names:
 
  .. code-block:: none
 
-    PolyData (0x7efc9da17dc8)
+    PolyData (0x7f5bac9ceb88)
       N Cells:      191
       N Points:     191
       X Bounds:     3.565e+05, 3.702e+05
@@ -83,7 +83,7 @@ Download sample data files and keep track of names:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.798 seconds)
+   **Total running time of the script:** ( 0 minutes  0.843 seconds)
 
 
 .. _sphx_glr_download_examples_ubc_read-grv-obs.py:
