+++ _sources/examples/gslib/sg_execution_times.rst.txt
@@ -5,7 +5,7 @@
 
 Computation times
 =================
-**00:49.400** total execution time for **examples_gslib** files:
+**00:48.082** total execution time for **examples_gslib** files:
 
-- **00:48.830**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
-- **00:00.570**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
+- **00:47.600**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
+- **00:00.482**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
