### 2019-06-27 #1451 @banesullivan
Latest version of the paper looks great!
