+++ PVPlugins/PVGeo_UBC_Tools.py
@@ -1,4 +1,4 @@
-paraview_plugin_version = '2.0.0'
+paraview_plugin_version = '2.0.1'
 # This is module to import. It provides VTKPythonAlgorithmBase, the base class
 # for all python-based vtkAlgorithm subclasses in VTK and decorators used to
 # 'register' the algorithm with ParaView along with information about UI.
