### 2019-06-28 #1451 @banesullivan
@leouieda - releases are being pushed on version `v2.0.1`

Zenodo: [![DOI](https://zenodo.org/badge/97628892.svg)](https://zenodo.org/badge/latestdoi/97628892)
