+++ _sources/examples/filters-general/many-slices-along-points.rst.txt
@@ -113,7 +113,7 @@ slices to what you prefer!
 
  .. code-block:: none
 
-    MultiBlock (0x7f5bac9f8828)
+    MultiBlock (0x7eff6603e588)
       N Blocks:     6
       X Bounds:     0.000, 250.000
       Y Bounds:     0.000, 250.000
@@ -143,7 +143,7 @@ slices to what you prefer!
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  3.713 seconds)
+   **Total running time of the script:** ( 0 minutes  3.780 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_many-slices-along-points.py:
