+++ _sources/content/PVGeo/index.rst.txt
@@ -4,7 +4,7 @@ About PVGeo
 * Author: Bane Sullivan
 * License: BSD-3-Clause
 * Copyright: 2018, Bane Sullivan
-* Version: 2.0.0
+* Version: 2.0.1
 
 
 
