+++ _sources/examples/filters-general/normalize-array.rst.txt
@@ -72,7 +72,7 @@ Create some input data. this can be any `vtkDataObject`
 
  .. code-block:: none
 
-    UniformGrid (0x7f5bac9f8a08)
+    UniformGrid (0x7eff6603e768)
       N Cells:      729
       N Points:     1000
       X Bounds:     0.000e+00, 9.000e+00
@@ -101,7 +101,7 @@ Create some input data. this can be any `vtkDataObject`
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  2.258 seconds)
+   **Total running time of the script:** ( 0 minutes  2.282 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_normalize-array.py:
