### 2019-06-29 #1451 @kyleniemeyer
Hi @banesullivan, I'm about to look over the paper to do some last checks, but first, can you please clean up the Zenodo archive metadata? The list of authors should match that of the paper, and right now you have a few extra people listed.
