### 2019-06-29 #1451 @kyleniemeyer
@banesullivan I just submitted a small PR that fixes a few citation commands in the paper, could you merge that? https://github.com/OpenGeoVis/PVGeo/pull/52

Also, is there a better reference for VTK.js? At minimum, some names or a group should be included as the author. I'm also wondering if they have a preferred citation.
