### 2019-06-29 d1a429a
Fixed a few citation commands in paper.md (#52)
