+++ _sources/examples/gslib/read-sgems-grid.rst.txt
@@ -46,7 +46,7 @@ Read SGeMS Grid file formats. The examples shown here are downloaded from
 
  .. code-block:: none
 
-    UniformGrid (0x7eff6489ad68)
+    UniformGrid (0x7f7b57384288)
       N Cells:      78000
       N Points:     157122
       X Bounds:     0.000e+00, 2.600e+02
@@ -146,7 +146,7 @@ Read SGeMS Grid file formats. The examples shown here are downloaded from
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  48.277 seconds)
+   **Total running time of the script:** ( 0 minutes  47.260 seconds)
 
 
 .. _sphx_glr_download_examples_gslib_read-sgems-grid.py:
