+++ _sources/examples/ubc/read-tensor-mesh.rst.txt
@@ -63,7 +63,7 @@ Read the mesh and model
 
  .. code-block:: none
 
-    RectilinearGrid (0x7eff648c5348)
+    RectilinearGrid (0x7f7b57f531c8)
       N Cells:      190440
       N Points:     200900
       X Bounds:     3.550e+05, 3.722e+05
@@ -127,7 +127,7 @@ Or threshold a data range
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  4.688 seconds)
+   **Total running time of the script:** ( 0 minutes  4.161 seconds)
 
 
 .. _sphx_glr_download_examples_ubc_read-tensor-mesh.py:
