+++ _sources/examples/gslib/sg_execution_times.rst.txt
@@ -5,7 +5,7 @@
 
 Computation times
 =================
-**00:49.308** total execution time for **examples_gslib** files:
+**00:47.756** total execution time for **examples_gslib** files:
 
-- **00:48.277**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
-- **00:01.030**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
+- **00:47.260**: :ref:`sphx_glr_examples_gslib_read-sgems-grid.py` (``read-sgems-grid.py``)
+- **00:00.496**: :ref:`sphx_glr_examples_gslib_read-point-set.py` (``read-point-set.py``)
