+++ _sources/examples/filters-general/rotate-points.rst.txt
@@ -62,7 +62,7 @@ Use the filter:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.727 seconds)
+   **Total running time of the script:** ( 0 minutes  0.691 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_rotate-points.py:
