### 2019-06-29 #1451 @banesullivan
@kyleniemeyer

> is there a better reference for VTK.js? At minimum, some names or a group should be included as the author. I'm also wondering if they have a preferred citation.

I'm not sure if there is another reference for VTK.js. Considering the VTK.js project's [README](https://github.com/Kitware/vtk-js/blob/master/README.md) states:
> vtk.js aims to be a subset of VTK

And points to the VTK Textbook reference that we currently have, maybe it would be fine to use the VTK reference?
