### 2019-06-29 #1451 @kyleniemeyer
It's good to cite software directly, in addition to a published work. So perhaps in the text, when you mention VTK.js, you can cite the textbook in addition to the VTK.js software. For the software citation, perhaps add `author = {{Kitware}}`, and then add the version you use and the year it was released. If that is the current version, you can just use "version 8.11.1" and 2019.
