### 2019-06-30 ef5cff9
Fix VTKjs reference
