+++ joss/paper.bib
@@ -12,7 +12,7 @@
     title={The paraview guide: a parallel visualization application},
     author={Ayachit, Utkarsh},
     year={2015},
-    publisher={Kitware, Inc.},
+    publisher={{Kitware, Inc.}},
 }
 
 
@@ -20,7 +20,7 @@
     title={The Visualization Toolkit (4th ed.)},
     author={Schroeder, Will J and Lorensen, Bill and Martin, Ken},
     year={2006},
-    publisher={Kitware},
+    publisher={{Kitware, Inc.}},
     isbn={978-1-930934-19-1}
 }
 
@@ -50,10 +50,11 @@
 }
 
 @misc{vtkjs,
-    title={VTK.js Documentation},
-    url={https://kitware.github.io/vtk-js/docs/},
-    note = {visited on 2018-08-01},
-    journal={ParaView},
+    title={VTK.js},
+    url={https://github.com/Kitware/vtk-js},
+    note = {v8.11.1 released on 2019-06-27},
+    author = {{Kitware, Inc.}},
+
 }
 
 @article{pyvista,
