### 2019-06-30 2cd06bd
Fix VTKjs reference
