+++ joss/paper.bib
@@ -52,7 +52,9 @@
 @misc{vtkjs,
     title={VTK.js},
     url={https://github.com/Kitware/vtk-js},
-    note = {v8.11.1 released on 2019-06-27},
+    year = {2019},
+    month = {June},
+    volume = {v8.11.1},
     author = {{Kitware, Inc.}},
 
 }
