+++ _sources/examples/gslib/read-point-set.rst.txt
@@ -45,7 +45,7 @@ Read GSLib point set file
 
  .. code-block:: none
 
-    PolyData (0x7f7b57f5e468)
+    PolyData (0x7f8a5203ad68)
       N Cells:      100
       N Points:     100
       X Bounds:     3.000e+00, 2.520e+02
@@ -72,7 +72,7 @@ Read GSLib point set file
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.496 seconds)
+   **Total running time of the script:** ( 0 minutes  0.501 seconds)
 
 
 .. _sphx_glr_download_examples_gslib_read-point-set.py:
