+++ _sources/examples/filters-general/extract-array.rst.txt
@@ -58,12 +58,12 @@ Construct the filter
 
  .. code-block:: none
 
-    vtkTable (0x23ac4b0)
+    vtkTable (0x3d9f810)
       Debug: Off
       Modified Time: 21530
       Reference Count: 2
       Registered Events: (none)
-      Information: 0x2368230
+      Information: 0x3ff0160
       Data Released: False
       Global Release Data: Off
       UpdateTime: 21542
@@ -103,7 +103,7 @@ Construct the filter
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.070 seconds)
+   **Total running time of the script:** ( 0 minutes  0.071 seconds)
 
 
 .. _sphx_glr_download_examples_filters-general_extract-array.py:
