+++ _sources/examples/grids/read-esri.rst.txt
@@ -58,7 +58,7 @@ Apply a filter to the DEM to have realistic topography
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  12.781 seconds)
+   **Total running time of the script:** ( 0 minutes  12.355 seconds)
 
 
 .. _sphx_glr_download_examples_grids_read-esri.py:
