### 2019-06-30 #1451 @kyleniemeyer
@banesullivan congratulations on your submission's publication in JOSS! Thanks to @gassmoeller and @Leguark for reviewing and @leouieda for editing.
