### 2019-06-30 #1451 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
